### 2025-01-30 #274 @editorialbot
[PRE REVIEW]: evolved: open software for teaching Evolutionary biology through virtual inquiries
**Submitting author:** @mjanuario (<a href="http://orcid.org/0000-0002-6480-7095">Matheus Januario</a>)
**Repository:** https://github.com/mjanuario/evolved
**Branch with paper.md** (empty if default branch): article
**Version:** V1.0
