### 2025-02-01 bd7b0f7
small fix in dev version text on README.Rmd
