+++ README.Rmd
@@ -35,7 +35,7 @@ install.packages("evolved")
 
 ### Development Version:
 
-The development version is the most complete and updated version of the package. It may contain datasets, functions, or documentation slightly out of CRAN's standards. The development version is hosted on [GitHub](https://github.com/mjanuario/evolved) and requires the `devtools` package for installation. If `devtools` is not already installed, you can install it first by uncommenting the line in the code below:
+The development version is the most complete and updated version of the package. It may contain datasets, functions, or documentation slightly out of CRAN's standards. The development version is hosted on [GitHub](https://github.com/mjanuario/evolved) and requires the `devtools` package for installation. If `devtools` is not already installed, you can install it first by running the code lines below:
 
 ```
 # Install devtools 
