### 2025-02-01 09855ad
small fix in dev-ver related text of README.md
