### 2025-06-24 #274 @labarba
The error says `branch name is incorrect`. I have manually edited the top comment to remove the branch (as it should be left empty for the default branch). Not sure if this will work...
