### 2025-06-24 #274 @labarba
@mjanuario – we can't find your `paper.md` file in the repository. See: https://openjournals.readthedocs.io/en/jose/submitting.html#what-should-my-paper-contain
