### 2025-07-03 c8228f6
Added files for JOSe review
