+++ paper.bib
@@ -1,0 +1,140 @@
+@article{BanchiBell:2008,
+  title={The many levels of inquiry.},
+  author={Banchi, Heather and Bell, Randy},
+  journal={Science \& Children},
+  volume={46},
+  number={2},
+  year={2008}
+}
+
+@article{Benson:2012,
+  title={GenBank},
+  author={Benson, Dennis A and Cavanaugh, Mark and Clark, Karen and Karsch-Mizrachi, Ilene and Lipman, David J and Ostell, James and Sayers, Eric W},
+  journal={Nucleic acids research},
+  volume={41},
+  number={D1},
+  pages={D36--D42},
+  year={2012},
+  publisher={Oxford University Press}
+}
+
+@article{Fitzjohn:2012,
+  title={Diversitree: comparative phylogenetic analyses of diversification in R},
+  author={FitzJohn, Richard G},
+  journal={Methods in Ecology and Evolution},
+  volume={3},
+  number={6},
+  pages={1084--1092},
+  year={2012},
+  publisher={Wiley Online Library}
+}
+
+@article{Januario:2024,
+  title={Package 'evolved'},
+  author={Januario, Matheus and Auler, Jennifer and Viol, Andressa and Rabosky, Daniel},
+  year={2024}
+}
+
+@article{Jetz:2012,
+  title={The global diversity of birds in space and time},
+  author={Jetz, Walter and Thomas, Gavin H and Joy, Jeffery B and Hartmann, Klaas and Mooers, Arne O},
+  journal={Nature},
+  volume={491},
+  number={7424},
+  pages={444--448},
+  year={2012},
+  publisher={Nature Publishing Group UK London}
+}
+
+@article{Jones:2009,
+  title={PanTHERIA: a species-level database of life history, ecology, and geography of extant and recently extinct mammals: Ecological Archives E090-184},
+  author={Jones, Kate E and Bielby, Jon and Cardillo, Marcel and Fritz, Susanne A and O'Dell, Justin and Orme, C David L and Safi, Kamran and Sechrest, Wes and Boakes, Elizabeth H and Carbone, Chris and others},
+  journal={Ecology},
+  volume={90},
+  number={9},
+  pages={2648--2648},
+  year={2009},
+  publisher={Wiley Online Library}
+}
+
+@article{Paradis:2019,
+  title={ape 5.0: an environment for modern phylogenetics and evolutionary analyses in R},
+  author={Paradis, Emmanuel and Schliep, Klaus},
+  journal={Bioinformatics},
+  volume={35},
+  number={3},
+  pages={526--528},
+  year={2019},
+  publisher={Oxford University Press}
+}
+
+@article{Pedaste:2015,
+  title={Phases of inquiry-based learning: Definitions and the inquiry cycle},
+  author={Pedaste, Margus and M{\"a}eots, Mario and Siiman, Leo A and De Jong, Ton and Van Riesen, Siswa AN and Kamp, Ellen T and Manoli, Constantinos C and Zacharia, Zacharias C and Tsourlidaki, Eleftheria},
+  journal={Educational research review},
+  volume={14},
+  pages={47--61},
+  year={2015},
+  publisher={Elsevier}
+}
+
+@article{R:2024,
+  title={{R: A language and environment for statistical computing}},
+  author={R Core Team},
+  journal={Foundation for Statistical Computing, Vienna, Austria},
+  year={2024}
+}
+
+@article{Rabosky:2014,
+  title={BAMM tools: an R package for the analysis of evolutionary dynamics on phylogenetic trees},
[…]
