+++ paper.md
@@ -1,0 +1,149 @@
+---
+title: 'evolved: open software for teaching Evolutionary biology through virtual inquiries'
+tags:
+  - R
+  - Ecology
+  - Evolution
+  - Inquiry-based learning
+authors:
+  - name: Matheus Januario
+    orcid: 0000-0002-6480-7095
+    affiliation: "1, 2" # (Multiple affiliations must be quoted)
+  - name: Jennifer P. Auler
+    orcid: 0000-0001-7576-9058
+    affiliation: 3
+  - name: Andressa L. Viol
+    affiliation: "1, 2"
+  - name: Daniel L. Rabosky
+    orcid: 0000-0002-7499-8251
+    affiliation: "1, 2"
+affiliations:
+ - name: Museum of Zoology, University of Michigan, Ann Arbor, MI, USA 
+   index: 1
+ - name: Department of Ecology and Evolutionary Biology, University of Michigan, Ann Arbor, MI, USA 
+   index: 2
+ - name: Programa de Pós-Graduação em Ecologia, Universidade de São Paulo, Rua do Matão, 321, Travessa 14, São Paulo, SP, Brazil
+   index: 3
+date: 19 December 2024
+bibliography: paper.bib
+
+---
+
+# Summary
+
+  `evolved` (**EVOL**utionary **V**irtual **ED**ucation) is an open-source R 
+  package [@R:2024] designed for graduate or advanced undergraduate courses in 
+  evolutionary biology. It provides tools for inquiry-based learning, where 
+  students engage in scientific practices to actively build knowledge
+  [@Pedaste:2015]. The package includes vignettes (tutorials) to facilitate 
+  classroom investigations (**Tables 1** & **2**). However, educators are encouraged 
+  to develop their own content modules depending on class context and/or learning 
+  objectives. Most of evolved's core functions are oriented towards either (i) 
+  developing student intuition for classic models in evolutionary biology 
+  using simulations, or (ii) analyzing (parameter estimation, hypothesis 
+  testing, visualization) datasets associated with key questions in 
+  evolutionary biology (**Tables 1** & **2**).  Importantly, these datasets correspond 
+  structurally (e.g., organization and data types) to real datasets used by 
+  scientists, and they are compatible with widely-used software for evolutionary 
+  analysis (e.g., `ape` - [@Paradis:2019]). 
+
+  `evolved`'s functions can be loosely classified into "high level" and "low level"
+  (**Table 1**), thus providing tools for inquiries that assume different levels 
+  of student independence (e.g., as categorized by @BanchiBell:2008 - see **Table 2**). High-level functions 
+  execute complex sequences of operations while hiding many details, thus making 
+  minimal assumptions of R proficiency and allowing users to focus on results. 
+  They are either optimized for performance, or work as a simplified interface 
+  ("wrappers") of simulators from standard scientific packages (e.g., functions 
+  from packages `ape` [@Paradis:2019], `diversitree` [@Fitzjohn:2012], or 
+  `phytools` [@Revell:2024]). These functions will likely work better in inquiries 
+  with more student freedom and a greater focus on analyzing data (**Table 2**). In 
+  contrast, "low level" functions focus on making explicit connections between 
+  every sequential operation and are thus more easily applicable to structured 
+  inquiries [@BanchiBell:2008]. They are also internally designed to be readable 
+  by students with intermediate skills in R, allowing further engagement with 
+  simulation parameters and outcomes.
+
+  The package is designed to reduce barriers to data- and simulation-driven inquiries 
+  and one vignette (`intro_R`) aims to introduce students/educators to R and its 
+  pedagogical capabilities. We do not provide answers to the investigation prompts 
+  in the vignettes, but educators can request answer keys from the corresponding 
+  author (MJ). 
+
[…]

[…]
