### 2025-07-03 877ebae
Del dir (paper branch)
