+++ R/WFDriftSim.R
@@ -1,152 +1,0 @@
-#' @importFrom  stats rbinom
-#' @importFrom  graphics lines
-#' @import grDevices
-NULL 
-
-#' Simulating generations of genetic drift in a Wright–Fisher (WF) population
-#'
-#' \code{WFDriftSim} simulates genetic drift of diploid Wright–Fisher 
-#' populations with a given effective population size through a certain number of 
-#' generations.
-#'
-#' @param Ne Number giving the effective population size of the population
-#' @param n.gen Number of generations to be simulated.
-#' @param p0 Initial frequency of a given allele. As the simulated organism is
-#' diploid, the other alleles frequency will be \code{1-(p0)}. Default value 
-#' is \code{0.5}.
-#' @param n.sim Number of simulations to be made. If decimals are inserted, 
-#' they will be rounded. Default value is \code{1}.
-#' @param plot.type Character indicating if simulations should be plotted as colored 
-#' lines. Each color represents a different population. If 
-#' \code{plot.type = "animate"} (default value) it animates each generation 
-#' individually. If \code{plot.type = "static"} it plots all lines rapidly. If 
-#' \code{plot.type = "none"} nothing is plotted.
-#' @param print.data Logical indicating whether all simulation results should be 
-#' returned as a \code{data.frame}. Default value is \code{FALSE}.
-#' @param knitr Logical indicating if plot is intended to show up in RMarkdown files made by the \code{Knitr} R package.
-#' 
-#' @return If \code{plot.type = "static"} or \code{"animate"}, plots the
-#' timeseries of all simulations, with each line+color referring to a 
-#' different simulation. Note that if many simulations (generally more 
-#' than 20) are simulated, colors might be cycled and different simulation 
-#' will have the same color. If \code{print.data = TRUE}, returns a 
-#' \code{data.frame} with the simulation results.
-#' 
-#' @details The effective population size (\code{Ne}) is strongly connected 
-#' with the rate of genetic drift (for details, see Waples, 2022).
-#' 
-#' @export WFDriftSim
-#' 
-#' @references 
-#' 
-#' Fisher RA (1922) On the dominance ratio. Proc. R. Soc. Edinb 42:321–341
-#' 
-#' Kimura M (1955) Solution of a process of random genetic drift with a 
-#' continuous model. PNAS–USA 41(3):144–150
-#' 
-#' Tran, T. D., Hofrichter, J., & Jost, J. (2013). An introduction to the 
-#' mathematical structure of the Wright–Fisher model of population genetics. 
-#' Theory in Biosciences, 132(2), 73-82. \[good for the historical review, 
-#' math can be challenging\]
-#' 
-#' Waples, R. S. (2022). What is Ne, anyway?. Journal of Heredity.
-#' 
-#' Wright S (1931) Evolution in Mendelian populations. Genetics 16:97–159
-#' 
-#' @author Matheus Januario, Dan Rabosky, Jennifer Auler
-#' 
-#' @examples
-#' \donttest{
-#' #Default values:
-#' WFDriftSim(Ne = 5, n.gen = 10, knitr = TRUE)
-#' 
-#' #A population which has already fixed one of the alleles:
-#' WFDriftSim(Ne = 5, n.gen = 10, p0=1, knitr = TRUE)
-#' 
-#' #Many populations::
-#' WFDriftSim(Ne = 5, n.gen = 10, p0=0.2, n.sim=10, knitr = TRUE)
-#' 
-#' ######## continuing a previous simulation:
[…]
