### 2025-07-03 fac8d5b
Del dir (paper branch)
