+++ data-raw/TO_DO.R
@@ -1,19 +1,0 @@
-# TO DO LIST:
-
-# 1) For CRAN release:
-# usethis::use_news_md() ok
-# usethis::use_cran_comments() (ok but to again in submition)
-# Update (aspirational) install instructions in README
-# Check that all exported functions have @returns and @examples
-# Check licensing of included files
-# Review https://github.com/DavisVaughan/extrachecks 
-
-# 2) create at least 1 testhat check per function, even 
-
-#-----
-
-# If updating readme with new vignettes/labs, please use devtools::build_readme() after.
-# update manuscript with new vignettes/labs
-
-#DONE#
-#######
