+++ data-raw/whale_body_size.R
@@ -1,3 +1,0 @@
-data_whales <- read.csv("./data-raw/whale_body_size.csv")
-
-usethis::use_data(data_whales, overwrite = TRUE)
