+++ data-raw/birds_spp.R
@@ -1,5 +1,0 @@
-birds <- read.csv("./data-raw/birds_spp.csv")
-
-birds_spp = birds$x
-
-usethis::use_data(birds_spp, overwrite = TRUE)
