+++ data-raw/all_fossil_data.R
@@ -1,11 +1,0 @@
-ammonoidea_fossil = read.csv(file = "data-raw/ammonoidea_fossil.csv")
-mammals_fossil = read.csv(file = "data-raw/mammals_fossil.csv")
-dinos_fossil = read.csv(file = "data-raw/dinos_fossil.csv")
-trilob_fossil = read.csv(file = "data-raw/trilob_fossil.csv")
-timeseries_fossil = read.csv(file = "data-raw/timeseries_fossil.csv")
-
-usethis::use_data(ammonoidea_fossil, overwrite = TRUE)
-usethis::use_data(mammals_fossil, overwrite = TRUE)
-usethis::use_data(dinos_fossil, overwrite = TRUE)
-usethis::use_data(trilob_fossil, overwrite = TRUE)
-usethis::use_data(timeseries_fossil, overwrite = TRUE)
