+++ data-raw/cytOxidase.txt
@@ -1,35 +1,0 @@
-
->cnidaria
------RWIFSTNHKDIGTLYLVFGLFSGMVGTALSMLIRLELSGPGTMFGDDHLYNVIVTAHAFVMIFFLVMPVLIGGFGNWFVPLFIGAPDMAFPRLNNLSFWLLPPALLLLLGSSLIEQGAGTGWTVYPPLSGPQTHSGGSVDMAIFSLHCAGASSIMGAINFITTIFNMRAPGMTMDKLPLFVWSVLITAFLLLLSLPVLAGAITMLLTDRNFNTTFFDPAGGGDPVLYQHLFWFFGHPEVYILILPAFGIISQIIPVFSSKNQIFGYLGMVYAQMGIAFLGFIVWAHHMFTIGMDVDTRAYFTAATMIIAVPTGIKIFSWLATMYGGKITFTTPMLFATGFIFLFTVGGLTGVILANGSLDVLLHDTYYVVAHFHYVLSLGAVFGMFAGFYFWFGKITGYAYNEIYGKIHFWVTFIGVNLTFFPQHFLGLAGMPRRYSDFPDSFATWNLVSSVGSLFSIVGVVWFIFTIYDAFVREEKFVSWGNTSEINTLEWIFNS-PPAHHTYNELPFIFYNNNK----------------
->snake
--MYITRWLFSTNHKDIGTLYLMFGAWSGLVGACLSILMRMELTQPGTLFGSDQIFNVLVTAHAFIMIFFMVMPIMIGGFGNWLIPLMLGTPDMAFPRMNNMSFWLLPPALLLLLSSSYIEAGAGTGWTVYPPLSGNLVHSGPSVDLAIFSLHLAGASSILGAINFITTCINMKPKSMPMFNIPLFVWSVMITAIMLLLALPVLAAAVTMLLTDRNLNTTFFDPCGGGDPVLFQHLFWFFGHPEVYILILPGFGIVSSIITFYTGKKNTFGYTSMIWAMMSIAILGFVVWAHHMFTVGLDIDSRAYFTAATMIIAIPTGIKVFGWLATLAGGHIKWQTPIYWALGFIFLFTVGGMTGIILANSSLDIVLHDTYYVVAHFHYVLSMGAVFAIMGGLTHWFPLFTGYSLNQTLTKTQFWVMFLGVNMTFFPQHFLGLSGMPRRYSDFPDAFTLWNTISSIGSTISLVAVLMSLFIVWEALTYKRKPTI--QLGKKTHIEWLHGT-PPPYHTHTEPTFMPNNTFAPIREYISYMQWPWPEK
->echinoderm
[…]
