+++ data-raw/doc_birds_spp.txt
@@ -1,26 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/data.R
-\docType{data}
-\name{birds_spp}
-\alias{birds_spp}
-\title{Birds species list}
-\format{
-\subsection{\code{birds_spp}}{
-
-A \code{vector} with 9993 bird species, taken originally from Jetz et al (2012)
-}
-}
-\source{ Original: Jetz, W., Thomas, G. H., Joy, J. B., Hartmann, K., & Mooers, A. O. (2012). The global diversity of birds in space and time. Nature, 491(7424), 444-448.
-
-Actual file donwloaded from https://vertlife.org/data/
-\if{html}{\out{
-<??>
-}}
-}
-\usage{
-birds_spp
-}
-\description{
-A vector containing the names of all extant brid species, following Jetz et al (2012) taxonomy
-}
-\keyword{datasets}
