### 2025-07-03 c351ed7
Del dir (paper branch)
