### 2025-07-03 5cb2164
Del dir (paper branch)
