+++ tests/testthat/test-WFDriftSim.R
@@ -1,3 +1,0 @@
-test_that("allel is fix", {
-  expect_equal(WFDriftSim(N = 5, n.gen = 30, p0=1, plot.type = "none", print.data = TRUE)[,30],1)
-})
