+++ tests/testthat/test-OneGenHWSim.R
@@ -1,3 +1,0 @@
-test_that("fixed allel", {
-  expect_equal(OneGenHWSim(p=1)[,1], rep(50, 100))
-})
