### 2025-07-03 64d71dd
Del dir (paper branch)
