+++ vignettes/deeptime_clocks.Rmd
@@ -1,491 +1,0 @@
----
-title: "Allele frequency change and selection"
-author: "Matheus Januario, Andressa Viol, and Daniel Rabosky"
-date: "Jan 2024"
-output: rmarkdown::html_vignette
-vignette: >
- %\VignetteIndexEntry{deeptime_clocks}
- %\VignetteEngine{knitr::rmarkdown}
- %\VignetteEncoding{UTF-8}
----
-
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE)
-options(rmarkdown.html_vignette.check_title = FALSE)
-```
-
-# Learning objectives
-
-1. Exploring sequence data
-2. A simple measure of genetic distance
-3. The Poisson correction
-4. Building a molecular clock
-5. Molecular clocks and inferences about deep time
-6. Jukes-Cantor correction
-7. The uncertainty of your molecular clock
-
-# Introduction
-
-```{r}
-#loading packages we will need today:
-library(ape)
-library(evolved)
-```
-
-We have encountered many processes that change allele frequency in a population, but today we will study the actual genetic differences among species. We will also learn how to use that genetic information to reconstruct the evolutionary past of lineages with the so-called "molecular clock". To do that, we will use _real data_ on molecules to estimate the divergence (in numbers of substitutions) between homologous proteins in different organisms.
-
-Homologous sequences, together with information from the fossil record, will be used to "calibrate" the rate at which amino acid substitutions occur during protein evolution. More specifically, given an  estimate of the amount of divergence, plus an estimate of the rate at which divergence occurs, we can estimate how much time has passed since divergence.
-
-\fbox{\begin{minipage}{48em}
-Question 1:\\
-
-Why is it important to make sure that the compared sequences are homologous?
-\end{minipage}}
-
-# Exploring sequence data
-
-We have assembled a set of protein sequences from the cytochrome oxidase 1 gene. This gene, often known as CO1 ("see-oh-one"), is a mitochondrial gene that plays a key role in cellular respiration (e.g., the primary aerobic pathway to energy or ATP generation). CO1 contains approximately 513 amino acids and has been used in many studies to construct phylogenetic trees and estimate divergence times between taxa.
-
-We manually downloaded full-length genes of a number of focal taxa from the public database [GenBank](https://www.ncbi.nlm.nih.gov/genbank/) (more general info  about GenBank can be found in Benson et al., 2012). We then *aligned* the sequences using [CLUSTAL](https://www.ebi.ac.uk/jdispatcher/msa/clustalo) (more info in Sievers et al., 2011), a widely-used sequence alignment tool.
-
-Because this dataset is included in our R package, we can access it by simply typing: 
-
-```{r}
-data("cytOxidase")
-```
-
-Now we can pull out the sequence of a single organism by typing:
-
-```{r, results = F}
-cytOxidase["human"]
-```
-
-
-\fbox{\begin{minipage}{48em}
-Question 2:\\
-
-When you run the above in your own console, you will see a sequence of letters. Given what you know so far about CO1, answer in one sentence: What do these letters represent? What is their biological role?
-\end{minipage}}
-
-\begin{center}
-
--- 
-
-\end{center}
-
-Now, let's explore our sequences a little.
-
-How many sequences do we have?
-
-```{r}
-length(cytOxidase)
-```
-
-What are the organisms in our dataset?
-
-```{r}
-names(cytOxidase)
-```
-
-How long is each sequence?
-
-```{r}
-unlist(lapply(cytOxidase, nchar))
-```
-
-You can access any particular sequence by typing:
-
-```{r}
-cytOxidase["platypus"]
-```
-
-You can compute the number of differences by typing:
-
-```{r}
-countSeqDiffs(cytOxidase, "human", "alligator")  
-```
-
[…]
