+++ vignettes/birthdeath_deeptime.Rmd
@@ -1,405 +1,0 @@
----
-title: "Birth-death models and deep time"
-author: "Matheus Januario, Andressa Viol, and Daniel Rabosky"
-date: "Jan 2024"
-output: rmarkdown::html_vignette
-vignette: >
- %\VignetteIndexEntry{birthdeath_deeptime}
- %\VignetteEngine{knitr::rmarkdown}
- %\VignetteEncoding{UTF-8}
----
-
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE)
-options(rmarkdown.html_vignette.check_title = FALSE)
-```
-
-# Learning objectives
-
-1. The birth-death (BD) model 
-2. Deterministic expectations of the birth-death model
-3. Estimation under the simple birth-death
-4. Effects of variation on the net diversification rate
-5. The birth-death model is a stochastic process
-6. Adding in extinction
-7. Age-richness models and empirical richness through time
-
-# Introduction
-
-Today we will work with the main mathematical model used in diversity dynamics: the birth-death model. The main idea of the model is simple: at any instant in time, species can (I) split into two, (II) go extinct, or (III) do nothing. The model can be fully described (i.e. parameterized) by the rate of speciation (the "births") and the rate of extinction (the "deaths").
-
-```{r}
-library(evolved)
-```
-
-# The birth-death (BD) model
-
-The birth-death model is a simple stochastic process that is widely used to quantitatively study diversification (speciation and extinction). It is used in other contexts as well, as in ecology or in physics.
-
-The birth-death model we will use here has two parameters: $S$, the _rate of speciation_ (the rate at which one species will split into two species), and $E$ the _rate of extinction_. Usually, these rates are "per capita" (or "per lineage"), meaning that the total rate of speciation for a clade of N species would be $N \times S$ and $N \times E$ for extinction. 
-
-The compound parameter $S - E$ is known as the net diversification rate: it is the rate at which new species are produced, minus the rate at which they go extinct. If $S - E > 0$, the net production of new species is positive and the clade tends to increase exponentially in diversity through time. When $S - E < 0$, the clade will lose species through time on average, meaning the clade will be strongly biased to go extinct.
-
-### Glossary:
-
-$S =$ speciation rate (per lineage)
-
-$E =$ extinction rate (per lineage)
-
-$S - E = R =$ net diversification rate
-
-$K = \frac{E}{S} =$ relative extinction rate (becomes important later)
-
-**pure-birth model:** a special type of birth-death model where only speciation happens (i.e., E = 0)
-
-# Deterministic expectations of the BD model
-
-The expected species richness under the birth-death model is:
-
-\begin{equation}
-N_t = N_0 * e^{t (S - E)}
-\end{equation}
-
-where $N_0$ is the starting number of species, $t$ is the elapsed time, and $N_t$ is the number of species at time $t$. This is just a simple exponential growth model that you've seen before (recall the intro to popgen vignette, where we discussed Malthusian growth), except now we are dealing with the birth and death rates of _species_ and not individuals.
-
-
-\fbox{\begin{minipage}{48em}
-Question 1:\\
-
-Are the relative contributions of speciation and extinction to net diversification distinguishable in this simple model?\\
-
-Hint: What is the expected richness through time curve for S = 10, E = 9, versus S = 1, E = 0?  Plot those 2 cases in a panel side by side on the same scale.
-\end{minipage}}
-
-\fbox{\begin{minipage}{48em}
-Question 2:\\
-
[…]

[…]
