### 2025-07-03 bc4ac26
Del file (paper branch)
