+++ DESCRIPTION
@@ -1,34 +1,0 @@
-Package: evolved
-Title: Open Software for Teaching Evolutionary Biology at Multiple Scales Through Virtual Inquiries
-Version: 1.0.0
-Authors@R: 
-    c(person("Matheus", "Januario", email="januarioml.eco@gmail.com", role = c("aut", "cph", "cre"),
-           comment = c(ORCID = "0000-0002-6480-7095")),
-      person("Jennifer", "Auler", email="auler.jennifer@gmail.com", role = c("aut"),
-           comment = c(ORCID = "0000-0001-7576-9058")),
-      person("Andressa", "Viol", email="aviol@umich.edu", role = c("aut")),
-      person("Daniel", "Rabosky", email="drabosky@umich.edu", role = c("aut")))
-Description: "Evolutionary Virtual Education" - 'evolved' - provides multiple tools to help educators (especially at the graduate level or in advanced undergraduate level courses) apply inquiry-based learning in general evolution classes. In particular, the tools provided include functions that simulate evolutionary processes (e.g., genetic drift, natural selection within a single locus) or concepts (e.g. Hardy-Weinberg equilibrium, phylogenetic distribution of traits). More than only simulating, the package also provides tools for students to analyze (e.g., measuring, testing, visualizing) datasets with characteristics that are common to many fields related to evolutionary biology. Importantly, the package is heavily oriented towards providing tools for inquiry-based learning - where students follow scientific practices to actively construct knowledge. For additional details, see package's vignettes.
-License: GPL (>= 3)
-Encoding: UTF-8
-Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.3.2
-Imports:
-  ape,
-  diversitree,
-  phytools
-Suggests: 
-    knitr,
-    rmarkdown,
-    testthat (>= 3.0.0),
-    LSD,
-    paleobioDB,
-    BAMMtools,
-    plot3D
-Config/testthat/edition: 3
-Depends: 
-    R (>= 2.10)
-LazyData: true
-VignetteBuilder: knitr
-URL: <https://github.com/Auler-J/evolved>
-BugReports: https://github.com/Auler-J/evolved/issues
