### 2025-07-03 578b969
Del file (paper branch)
