+++ NAMESPACE
@@ -1,36 +1,0 @@
-# Generated by roxygen2: do not edit by hand
-
-S3method(head,ProteinSeq)
-S3method(print,ProteinSeq)
-S3method(summary,ProteinSeq)
-S3method(tail,ProteinSeq)
-export(NatSelSim)
-export(OneGenHWSim)
-export(WFDriftSim)
-export(calcFossilDivTT)
-export(checkAndFixUltrametric)
-export(countSeqDiffs)
-export(estimateSpeciation)
-export(fitCRBD)
-export(is.ProteinSeq)
-export(lttPlot)
-export(plotNatSel)
-export(plotPaintedWhales)
-export(plotProteinSeq)
-export(plotRawFossilOccs)
-export(plotWFDrift)
-export(simulateBirthDeathRich)
-export(simulateTree)
-import(grDevices)
-import(graphics)
-importFrom(ape,is.binary.phylo)
-importFrom(graphics,lines)
-importFrom(graphics,par)
-importFrom(graphics,plot)
-importFrom(stats,aggregate)
-importFrom(stats,rbinom)
-importFrom(stats,rgeom)
-importFrom(stats,runif)
-importFrom(utils,data)
-importFrom(utils,head)
-importFrom(utils,tail)
