### 2025-07-03 ee5c85a
Del file (paper branch)
