+++ NEWS.md
@@ -1,14 +1,0 @@
-# What's New
-
-## Version: 1.0.0
-Date: Mar/2024
-
-Official release.
-* Added vignettes to help teachers use package's tools.
-* Added generic functions to "ProteinSeq" class.
-
-## Version: 0.0.0.5000
-Date: Feb/2023
-
-Test release of the package.
-* All main functions are implemented.
