### 2025-07-03 7c99924
Del file (paper branch)
