+++ README.Rmd
@@ -1,431 +1,0 @@
----
-title: 'Instructions on EvolVEd, R, RStudio, RMarkdown'
-author: "Matheus Januario, Jenniefer Auler, Andressa Viol, and Daniel Rabosky"
-date: "Nov/2024"
-output: 
-  github_document:
-    toc: true
-    toc_depth: 2 # Adjust the depth of headers included in the TOC if needed
-citecolor: blue
-urlcolor: blue
----
-
-```{r setup, include=FALSE}
-#this chunk of code is purposedly not included in the PDF (note the header of 
-# the code chunk)
-knitr::opts_chunk$set(echo = TRUE)
-#setwd("~/Dropbox (University of Michigan)/devel/EEB_391_516_labs_2023/L00/")
-```
-
-# EvolVEd
-
-`evolved` (EVOLutionary Virtual EDucation) is an open-source R package designed for graduate or advanced undergraduate courses in evolutionary biology. It emphasizes tools for inquiry-based learning, where students engage in scientific practices to actively build knowledge (Pedaste et al., 2015). The package includes `vignettes` (tutorials) to facilitate classroom investigations. However, educators are encouraged to develop their own content modules depending on class context and/or learning objectives. Most of evolved‘s core functions are oriented towards either ($i$) developing student intuition for classic models in evolutionary biology using simulations, or ($ii$) analyzing (parameter estimation, hypothesis testing, visualization) datasets associated with key questions in evolutionary biology.
-
-## Installing EvolVEd
-
-The EvolVEd package is available in two versions:
-
-### Stable Version:
-
-The stable version represents the "error-proof" version of the package. It may have reduced features due to CRAN's constraints (e.g., dataset size). The stable version can be downloaded directly from the [The Comprehensive R Archive Network](https://cran.r-project.org/web/packages/evolved/index.html). A simpler way to install this version is to run the following command in your R terminal:
-
-```
-install.packages("evolved")
-```
-
-### Development Version:
-
-The development version is the most complete and updated version of the package. It may contain datasets, functions, or documentation slightly out of CRAN's standards. The development version is hosted on [GitHub](https://github.com/mjanuario/evolved) and requires the `devtools` package for installation. If `devtools` is not already installed, you can install it first by running the code lines below:
-
-```
-# Install devtools 
-install.packages("devtools")
-
-# Then, install the development version of EvolVEd:
-devtools::install_github("mjanuario/evolved")
-```
-
-#### Features in the development version that are absent in the stable version:
-
-* Datasets:
-  * `mammals_spp` (Extant species list of mammals)
-  * `ammonoidea_fossil` (Fossil occurrences of ammonoids
-  * `mammals_fossil` (Fossil occurrences of mammals)
-  * `trilob_fossil` (Fossil occurrences of trilobites)
-
-## Vignettes (Tutorials)
-
-To view the vignettes, run the following code:
-
-```
-vignette("vignette_name", package = "evolved")
-```
-
-With `"vignette_name"` being one of the names below. Each vignette covers the following topics, organized from basic to advanced:
-
-#### 1. Intro to R (name: `intro_r`)
-* Basic R syntax and coding (objects, vector calculations, etc.)
-* Plotting and annotation functions
-* Overview of key object classes used in the vignettes
-
-#### 2. Introduction to Population Genetics (name: `popgen_intro`)
-* Simple mathematical notation
-* Probability of independent events and random number generation
-* Malthusian growth and Mendelian genetics
-* Hardy-Weinberg Equilibrium (HWE) at a single locus
-* Heterozygosity, deleterious alleles, and mutation
-* Genetics at multiple loci and DNA variation
-
[…]
