### 2025-07-03 #274 @mjanuario
Hi @labarba ! Thank you for checking. The paper.md file is located in the article branch of the repository:

https://github.com/mjanuario/evolved/tree/article

Please let me know if there's anything else I can clarify!

Best regards,
Matheus
