### 2025-07-03 #274 @mjanuario
Hello Dr Barba,

I have modified the branch for my article submitted for JOSE
<https://github.com/openjournals/jose-reviews/issues/274#issuecomment-3001809759>.
Thank you very much for flagging my mistake!

Cheers,
Matheus

On Tue, Jun 24, 2025 at 5:07 PM Lorena A. Barba ***@***.***>
wrote:

> *labarba* left a comment (openjournals/jose-reviews#274)
> <https://github.com/openjournals/jose-reviews/issues/274#issuecomment-3001863293>
>
> @mjanuario <https://github.com/mjanuario> – we can't find your paper.md
> file in the repository. See:
> https://openjournals.readthedocs.io/en/jose/submitting.html#what-should-my-paper-contain
>
