### 2025-07-07 #274 @ethanwhite
Hi @mjanuario! I'll be serving as the editor on this paper. I've given everything a preliminary look through and it all looks good. Thanks for providing a list of recommended reviewers. I'll start inviting folks and we'll get the review process going. Looking forward to working with you on this!
