### 2025-07-07 #274 @ethanwhite
👋 @K8hertweck @pschloss @fmichonneau - would you be willing to review this submission for JOSE? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://openjournals.readthedocs.io/en/jose/review_criteria.html
