### 2025-08-05 #300 @editorialbot
[REVIEW]: evolved: open software for teaching Evolutionary biology through virtual inquiries
**Submitting author:** @mjanuario (<a href="http://orcid.org/0000-0002-6480-7095">Matheus Januario</a>)
**Repository:** https://github.com/mjanuario/evolved
**Branch with paper.md** (empty if default branch): article
**Version:** 1.0.23
**Paper kind:** software
