### 2025-08-05 #300 @ethanwhite
@mjanuario - Can you please add DOI's to the bibliography file when you get a minute and then ping me when that's done so I can rerun the references check. Most of them are listed above. In addition, PanTHERIA has an associated publication with a DOI, and all R packages now have DOIs so you can add one for BAMM tools as well.
