### 2025-08-05 a9b62b9
Update paper.bib
add ref DOIs
