+++ paper/paper.bib
@@ -5,6 +5,7 @@
   volume={46},
   number={2},
   year={2008}
+  doi={10.2505/3/sc08_046_02}
 }
 
 @article{Benson:2012,
@@ -16,6 +17,7 @@
   pages={D36--D42},
   year={2012},
   publisher={Oxford University Press}
+  doi={10.1093/nar/gkw1070}
 }
 
 @article{Fitzjohn:2012,
@@ -27,12 +29,14 @@
   pages={1084--1092},
   year={2012},
   publisher={Wiley Online Library}
+  doi={10.1111/j.2041-210X.2012.00234.x}
 }
 
 @article{Januario:2024,
   title={Package 'evolved'},
   author={Januario, Matheus and Auler, Jennifer and Viol, Andressa and Rabosky, Daniel},
   year={2024}
+  doi={10.32614/CRAN.package.evolved}
 }
 
 @article{Jetz:2012,
@@ -44,6 +48,7 @@
   pages={444--448},
   year={2012},
   publisher={Nature Publishing Group UK London}
+  doi={10.1038/nature11631}
 }
 
 @article{Jones:2009,
@@ -55,6 +60,7 @@
   pages={2648--2648},
   year={2009},
   publisher={Wiley Online Library}
+  doi={10.1890/08-1494.1}
 }
 
 @article{Paradis:2019,
@@ -66,6 +72,7 @@
   pages={526--528},
   year={2019},
   publisher={Oxford University Press}
+  doi={10.1093/bioinformatics/bty633}
 }
 
 @article{Pedaste:2015,
@@ -76,6 +83,7 @@
   pages={47--61},
   year={2015},
   publisher={Elsevier}
+  doi={10.1016/j.edurev.2015.02.003}
 }
 
 @article{R:2024,
@@ -83,6 +91,7 @@
   author={R Core Team},
   journal={Foundation for Statistical Computing, Vienna, Austria},
   year={2024}
+  url = {https://www.R-project.org/}
 }
 
 @article{Rabosky:2014,
@@ -94,6 +103,7 @@
   pages={701--707},
   year={2014},
   publisher={Wiley Online Library}
+  doi={10.1111/2041-210X.12199}
 }
 
 @article{Rabosky:2021,
@@ -105,6 +115,7 @@
   pages={2945},
   year={2021},
   publisher={Nature Publishing Group UK London}
+  doi={10.1038/s41467-021-23307-5}
 }
 
 @article{Revell:2024,
@@ -115,6 +126,7 @@
   pages={e16505},
   year={2024},
   publisher={PeerJ Inc.}
+  doi={10.7717/peerj.16505}
 }
 
 @article{Steeman:2009,
@@ -126,6 +138,7 @@
   pages={573--585},
   year={2009},
   publisher={Oxford University Press}
+  doi={10.1093/sysbio/syp060}
 }
 
 @article{Upham:2019,
@@ -137,4 +150,5 @@
   pages={e3000494},
   year={2019},
   publisher={Public Library of Science San Francisco, CA USA}
+  doi={10.1371/journal.pbio.3000494}
 }
