### 2025-08-05 #300 @mjanuario
@ethanwhite  Thanks for the heads up! Just finished adding the DOIs to the bibliography file. Let me know if anything else is missing on my end!
