### 2025-08-05 #300 @ethanwhite
Thanks @mjanuario! It looks like maybe there's a typo in the bibtex now. Let me know if you can't track it down and I can take a look.
