### 2025-08-05 32e25ba
Update paper.bib
fix comma typo on bibtex
