+++ paper/paper.bib
@@ -4,7 +4,7 @@
   journal={Science \& Children},
   volume={46},
   number={2},
-  year={2008}
+  year={2008},
   doi={10.2505/3/sc08_046_02}
 }
 
@@ -16,7 +16,7 @@
   number={D1},
   pages={D36--D42},
   year={2012},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
   doi={10.1093/nar/gkw1070}
 }
 
@@ -28,14 +28,14 @@
   number={6},
   pages={1084--1092},
   year={2012},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   doi={10.1111/j.2041-210X.2012.00234.x}
 }
 
 @article{Januario:2024,
   title={Package 'evolved'},
   author={Januario, Matheus and Auler, Jennifer and Viol, Andressa and Rabosky, Daniel},
-  year={2024}
+  year={2024},
   doi={10.32614/CRAN.package.evolved}
 }
 
@@ -47,7 +47,7 @@
   number={7424},
   pages={444--448},
   year={2012},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
   doi={10.1038/nature11631}
 }
 
@@ -59,7 +59,7 @@
   number={9},
   pages={2648--2648},
   year={2009},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   doi={10.1890/08-1494.1}
 }
 
@@ -71,7 +71,7 @@
   number={3},
   pages={526--528},
   year={2019},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
   doi={10.1093/bioinformatics/bty633}
 }
 
@@ -82,7 +82,7 @@
   volume={14},
   pages={47--61},
   year={2015},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi={10.1016/j.edurev.2015.02.003}
 }
 
@@ -90,7 +90,7 @@
   title={{R: A language and environment for statistical computing}},
   author={R Core Team},
   journal={Foundation for Statistical Computing, Vienna, Austria},
-  year={2024}
+  year={2024},
   url = {https://www.R-project.org/}
 }
 
@@ -102,7 +102,7 @@
   number={7},
   pages={701--707},
   year={2014},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   doi={10.1111/2041-210X.12199}
 }
 
@@ -114,7 +114,7 @@
   number={1},
   pages={2945},
   year={2021},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
   doi={10.1038/s41467-021-23307-5}
 }
 
@@ -125,7 +125,7 @@
   volume={12},
   pages={e16505},
   year={2024},
-  publisher={PeerJ Inc.}
+  publisher={PeerJ Inc.},
   doi={10.7717/peerj.16505}
 }
 
@@ -137,7 +137,7 @@
   number={6},
   pages={573--585},
   year={2009},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
   doi={10.1093/sysbio/syp060}
 }
 
@@ -149,6 +149,6 @@
   number={12},
   pages={e3000494},
   year={2019},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
   doi={10.1371/journal.pbio.3000494}
 }
