### 2025-08-05 17e15f2
Update paper.bib
try to fix DOI for B&B2008
