+++ paper/paper.bib
@@ -5,7 +5,7 @@
   volume={46},
   number={2},
   year={2008},
-  doi={10.2505/3/sc08_046_02}
+  url = {http://www.jstor.org/stable/43174976.}
 }
 
 @article{Benson:2012,
@@ -91,7 +91,8 @@
   author={R Core Team},
   journal={Foundation for Statistical Computing, Vienna, Austria},
   year={2024},
-  url = {https://www.R-project.org/}
+  url = {https://www.R-project.org/},
+  doi = {10.32614/r.manuals}
 }
 
 @article{Rabosky:2014,
