### 2025-08-05 #300 @mjanuario
Hi @ethanwhite I think there's no official DOI for this paper (Banchi & Bell 2008). I also checked the journal's issue on their website and the do not list a DOI there. How should I proceed?
