### 2025-08-05 #300 @gvdr
- [ ] **Version:** Does the release version given match the GitHub release?
