### 2025-08-06 #300 @ethanwhite
> Hi [@ethanwhite](https://github.com/ethanwhite) I think there's no official DOI for this paper (Banchi & Bell 2008). I also checked the journal's issue on their website and the do not list a DOI there. How should I proceed?

Yep, that happens and it's no problem. It just goes in without a DOI as you have it now. Thanks so much for getting that all taken care of so quickly!
