### 2025-08-18 #300 @ethanwhite
👋 @fmichonneau, @gvdr. Thanks again for agreeing to review this paper. I see that @gvdr's been making great progress on the checklist - much appreciated! Since we're now two weeks into the review process, I just wanted to check in and see if either of you have any questions or have run into any issues that I can help out with as you work on your reviews.
