### 2025-08-26 #300 @fmichonneau
- [x] **Version:** Does the release version given match the GitHub release?

## Suggestions

Below are some suggestions to improve the paper and the content of the repository. Except for adding the Community Guidelines which are required by JOSE, the others are left at the appreciation of the authors.

* Update the GitHub description, it's helpful to have some information there.

* Remove the `.DS_Store` files committed to the repository
* Regarding the "Version" in the checklist, I don't see the version number listed in the paper and the GitHub repository doesn't use the "Releases" feature (see this [chapter](https://r-pkgs.org/release.html#sec-release-post-acceptance) from R Packages book)  to have access to the version history of the package besides looking at the `NEWS.md` file.
* I would suggest building a `pkgdown` [website](https://r-pkgs.org/website.html) to go with the package (that can be hosted directly on GitHub) to be able to make it easier to browse the content of the vignettes, and make the work that went into this package more discoverable.
* The test suite is rather limited, but given that the package mostly rely on functions provided by others, it's probably OK here.
* There are no community guidelines found in the README.
