### 2025-08-30 #300 @fmichonneau
@ethanwhite I completed my initial review.
@mjanuario It's an interesting paper and I can see that the package provides a helpful way of teaching evolutionary biology. Don't hesitate to let me know if you have any comments or questions about my suggestions.
