### 2025-09-05 5fdd8b0
Update README.md
changed caps lokc in pkg name
