### 2025-09-05 ff30cbf
Update DESCRIPTION
typo fix
