### 2025-09-26 #300 @ethanwhite
@gvdr - just checking in one more time to see if you are able to finish your review

@mjanuario - apologies for the delay. Please go ahead and address @fmichonneau's suggestions when you have time while I discuss whether or not to replace the other reviewer with the Editor in Chief
