### 2025-10-10 #300 @ethanwhite
- [ ] **Version:** Does the release version given match the GitHub release?
