### 2025-10-10 #300 @ethanwhite
@mjanuario Nice work on this valuable contribution! I agree with all of @fmichonneau's suggestions. I see you've already addressed some of them by updatign the GitHub description, removing the `.DS _Store` file, and adding a GitHub release. Thanks! That leaves the following three points that I'll provide my take on here:

1. I agree with @fmichonneau that a pkgdown website would substantially expand the use of the vignettes that are core to this material. pkgdown is designed to make it easy to host this website directly through GitHub. See https://pkgdown.r-lib.org/ and let me know if you have any questions about if/how to proceed.
2. The test suite does feel thin. I agree with @fmichonneau that you probably don't have to have more unit tests, but I think testing the vignette builds would be valuable. I think if you add the pkgdown website this would generally satisfy this requirement, though ideally I'd recommend testing the package build with vignette's using GitHub Actions so that you know if the vignette's break before merging and having the site build fail.
3. You should add a CONTRIBUTING.md file to your repository describing how others can contribute. This file can be short and just point people to where they need to go to do basic things. E.g., https://github.com/weecology/forecasting-course/blob/main/CONTRIBUTING.md

Take a look through @fmichonneau's review and these followups on it and let me know if you have any questions or disagree with any of the remaining recommendations. Once you've made these changes we should be all set to get this paper into JOSE!
