### 2025-10-21 1059c4e
a simple yet funcional version of pkgdown
