+++ README.Rmd
@@ -106,6 +106,11 @@ With `"vignette_name"` being one of the names below. Each vignette covers the fo
 * Estimating diversification rates under pure birth and birth-death models
 * Factors influencing speciation and extinction
 
+## Community guidelines and suggesting changes
+
+Questions and support requests should be sent to Matheus Januario's email (januarioml.eco [at] gmail.com). 
+Bug & issue reports, suggestions, improvements, or code additions should be pushed through the same GitHub repository. Emailing Matheus Januario is also ok.
+
 # If you never used R before...
 
 If you’re new to R, this "pre-vignette" tutorial serves as an introduction to using `evolved`.  It was largely inspired by the first tutorials in EEB 485 and EEB 391/516 (Population and Community Ecology & Evolutionary processes and Macroevolution), but it is extended and modified to work in our context. This document is meant to serve as a first contact with the tools that will be of major importance in our vignettes with `evovled`. Nothing related to this tutorial is required to use the R package, but we will be using all the tools explained here, so please make sure you have completed it before then (or to be genrlly familiar with R).
