### 2025-11-28 e1beab3
knitted README
