### 2025-11-28 0060c9d
created install_r vignette
