+++ README.Rmd
@@ -105,327 +105,3 @@ With `"vignette_name"` being one of the names below. Each vignette covers the fo
 #### 8. Phylogenies and Birth-Death Models (name: `birthdeath_phylogenies`)
 * Estimating diversification rates under pure birth and birth-death models
 * Factors influencing speciation and extinction
-
-# If you never used R before...
-
-If you’re new to R, this "pre-vignette" tutorial serves as an introduction to using `evolved`.  It was largely inspired by the first tutorials in EEB 485 and EEB 391/516 (Population and Community Ecology & Evolutionary processes and Macroevolution), but it is extended and modified to work in our context. This document is meant to serve as a first contact with the tools that will be of major importance in our vignettes with `evovled`. Nothing related to this tutorial is required to use the R package, but we will be using all the tools explained here, so please make sure you have completed it before then (or to be genrlly familiar with R).
-
-This tutorial will be somewhat long, and if this is your first time using R, it will also be a bit confusing. This is because a reasonable part of what you will see here is pretty abstract. We ask you to trust the learning process and keep reading until the end. We highly encourage you to reach out to other peers more knowledgeable in R, if needed. We know that different folks come with very different levels of comfort with R -- which means you will learn a lot with your colleagues, however comfortable you may currently be with R.
-
-## What is R and how can I install it?
-
-R is (1) an open source software, (2) a programming language, and (3) a programming environment. The first one means it is a computer program you can freely download, use, or modify to your own needs. The second means it is a language that allows humans and machines to communicate. The third means it gives you everything you need to create your own programs under a specific language. Scientists, especially in the biological sciences, widely use R to run simulations and analyze their datasets, so we will do this too.
-
-If you’re using your own computer and have not yet installed R, you can download and install it for free by pasting this in your web browser:
-
-<http://cran.us.r-project.org/>
-
-After you download it, install it as you would do with any other program -- the installer program will direct you though the major steps. When you open R, you should see something close to Figure 1.
-
-Please download and install R before progressing to install RStudio.
-
-```{r, out.width='100%', fig.align='center', fig.cap='Example of recently-opened R on an Apple computer. The interface looks slightly different on a Windows computer, but the main features are the same. Note that depending on the details, you may or may not see the same warnings and errors as the screenshot.', echo=FALSE}
-knitr::include_graphics('README_images/R_screenshot.png')
-```
-
-## What is RStudio and how can I install it?
-
-RStudio is a development environment for R. As a very coarse simplification, you can think of RStudio as a complement to R that makes some tasks easier -- a "shell" enveloping R that makes some communications between you and R easier. However, many would say that RStudio is not much of a helpful resource at all. If you never worked with R before, you are encouraged to try both RStudio and just "plain" R separately and make your own conclusion as to which makes you more comfortable. You are not required to use RStudio to use `evolved`, but you can use it if you think it will help your learning process (you are required to use R, though).
-
-Here in this class we will use programming as a learning tool, but programming is also a skill that will develop your quantitative ability and will make you a better scientist. If you really want to train this skill and get really good at it, we highly recommend that at some point of your career, you get used to programming just with R -- and you are invited to use this class as one opportunity to do that. However, if your intention is to use R only occasionally, it might be worth jumping straight to RStudio, at the cost of being limited by a series of details and constraints that RStudio has.
-
[…]
