+++ README.md
@@ -1,22 +1,13 @@
-Instructions on evolved, R, RStudio, RMarkdown
+Instructions on EvolVEd, R, RStudio, RMarkdown
 ================
 Matheus Januario, Jenniefer Auler, Andressa Viol, and Daniel Rabosky
 Nov/2024
 
-- [evolved](#evolved)
-  - [Installing evolved](#installing-evolved)
+- [EvolVEd](#evolved)
+  - [Installing EvolVEd](#installing-evolved)
   - [Vignettes (Tutorials)](#vignettes-tutorials)
-- [If you never used R before…](#if-you-never-used-r-before)
-  - [What is R and how can I install
-    it?](#what-is-r-and-how-can-i-install-it)
-  - [What is RStudio and how can I install
-    it?](#what-is-rstudio-and-how-can-i-install-it)
-  - [Starting R and RStudio:](#starting-r-and-rstudio)
-  - [What is RMarkdown and how can I install it
-    (optional)?](#what-is-rmarkdown-and-how-can-i-install-it-optional)
-  - [Working Routine in R](#working-routine-in-r)
 
-# evolved
+# EvolVEd
 
 `evolved` (EVOLutionary Virtual EDucation) is an open-source R package
 designed for graduate or advanced undergraduate courses in evolutionary
@@ -31,37 +22,44 @@ biology using simulations, or ($ii$) analyzing (parameter estimation,
 hypothesis testing, visualization) datasets associated with key
 questions in evolutionary biology.
 
-## Installing evolved
+## Installing EvolVEd
 
-The evolved package is available in two versions:
+The EvolVEd package is available in two versions:
 
 ### Stable Version:
 
-The stable version represents the "error-proof" version of the package. It may have reduced features due to CRAN's constraints (e.g., dataset size). The stable version can be downloaded directly from the [The Comprehensive R Archive Network](https://cran.r-project.org/web/packages/evolved/index.html). A simpler way to install this version is to run the following command in your R terminal:
+The stable version represents the “error-proof” version of the package.
+It may have reduced features due to CRAN’s constraints (e.g., dataset
+size). The stable version can be downloaded directly from the [The
+Comprehensive R Archive
+Network](https://cran.r-project.org/web/packages/evolved/index.html). A
+simpler way to install this version is to run the following command in
+your R terminal:
 
-```
-install.packages("evolved")
-```
+    install.packages("evolved")
 
 ### Development Version:
 
-The development version is the most complete and updated version of the package. It may contain datasets, functions, or documentation slightly out of CRAN's standards. The development version is hosted on [GitHub](https://github.com/mjanuario/evolved) and requires the `devtools` package for installation. If `devtools` is not already installed, you can install it first by running the code lines below:
+The development version is the most complete and updated version of the
+package. It may contain datasets, functions, or documentation slightly
+out of CRAN’s standards. The development version is hosted on
+[GitHub](https://github.com/mjanuario/evolved) and requires the
+`devtools` package for installation. If `devtools` is not already
+installed, you can install it first by running the code lines below:
 
-```
-# Install devtools 
-install.packages("devtools")
+    # Install devtools 
+    install.packages("devtools")
 
-# Then, install the development version of evolved:
-devtools::install_github("mjanuario/evolved")
-```
+    # Then, install the development version of EvolVEd:
+    devtools::install_github("mjanuario/evolved")
 
 #### Features in the development version that are absent in the stable version:
 
-* Datasets:
-  * `mammals_spp` (Extant species list of mammals)
-  * `ammonoidea_fossil` (Fossil occurrences of ammonoids
[…]
