### 2025-11-28 1554267
build install_r vignette
