+++ vignettes/install_r.Rmd
@@ -33,7 +33,7 @@ After you download it, install it as you would do with any other program -- the 
 Please download and install R before progressing to install RStudio.
 
 ```{r, out.width='100%', fig.align='center', fig.cap='Example of recently-opened R on an Apple computer. The interface looks slightly different on a Windows computer, but the main features are the same. Note that depending on the details, you may or may not see the same warnings and errors as the screenshot.', echo=FALSE}
-knitr::include_graphics('README_images/R_screenshot.png')
+knitr::include_graphics('install_r_images/R_screenshot.png')
 ```
 
 ## What is RStudio and how can I install it?
@@ -55,7 +55,7 @@ If you intend to use RStudio, please download and install it before progressing 
 When you start R, you should see something like figure 1. If you start RStudio instead, it will open R inside of the RStudio environment, and should look like figure 2.
 
 ```{r, out.width='100%', fig.align='center', fig.cap='Example of recently-opened Rstudio. Note that the appearance of RStudio is customizable, so your interface might not look the same as the RStudio your instructors or your fellow students have', echo=FALSE}
-knitr::include_graphics('README_images/RStudio_screenshot.png')
+knitr::include_graphics('install_r_images/RStudio_screenshot.png')
 ```
 
 Just looking at figures 1 and 2, it is easy to see that RStudio gives a lot more information. We will briefly walk through the items shown in figure 2 to familiarize you with each quadrant in the RStudio environment. Note that in both figures we are not showing the upper part of the screen (containing the tabs "File", "Edit", "Code", "View", etc.), but sometimes you will use those options too.
@@ -89,7 +89,7 @@ In this short tutorial we will help you to navigate through your files. You can 
 We start by revising what the "Desktop" of your computer is -- the standard folder most computer operating systems (i.e., Windows, MacOS, Linux...) have. It is the first "place" your computer puts you in when you turn it on, seen in the screen shot below, which is my Desktop:
 
 ```{r, echo=F, out.width="100%"}
-knitr::include_graphics('README_images/set_dir_fig00.png')
+knitr::include_graphics('install_r_images/set_dir_fig00.png')
 ```
 
 \
@@ -101,7 +101,7 @@ If you are not comfortable with coding or R, we suggest you create a folder for 
 The working directory is the folder in your computer that R will navigate to during a session. Any time you import files or write files to a folder, R will automatically use the folder that is set as the working directory. To set the working directory, open R and click, in the upper bar, on _Session > Set Working Directory > Choose Directory_ as in the figure below (notice that my .Rmd file is on the left panel and my R console, different from the RStudio default, is on the right -- notice also that I minimized all other panels in my RStudio):
 
 ```{r, echo=F, out.width="100%"}
-knitr::include_graphics('README_images/set_dir_fig0.png')
+knitr::include_graphics('install_r_images/set_dir_fig0.png')
 ```
 
 \
@@ -109,13 +109,13 @@ knitr::include_graphics('README_images/set_dir_fig0.png')
 After clicking on _Choose Directory_, you should see something like the screenshot below:
 
 ```{r, echo=F, out.width="100%"}
-knitr::include_graphics('README_images/set_dir_fig1.png')
+knitr::include_graphics('install_r_images/set_dir_fig1.png')
 ```
 
[…]
