### 2025-11-28 bc8fc89
update pkgdown theme and vignettes organization
