### 2025-11-28 c13ceb1
Add install r vignette in readme
