+++ README.md
@@ -6,20 +6,6 @@ Nov/2024
 - [EvolVEd](#evolved)
   - [Installing EvolVEd](#installing-evolved)
   - [Vignettes (Tutorials)](#vignettes-tutorials)
-<<<<<<< HEAD
-  - [Community guidelines and suggesting
-    changes](#community-guidelines-and-suggesting-changes)
-- [If you never used R before…](#if-you-never-used-r-before)
-  - [What is R and how can I install
-    it?](#what-is-r-and-how-can-i-install-it)
-  - [What is RStudio and how can I install
-    it?](#what-is-rstudio-and-how-can-i-install-it)
-  - [Starting R and RStudio:](#starting-r-and-rstudio)
-  - [Working Routine in R](#working-routine-in-r)
-- [What is RMarkdown and how can I install it
-  (optional)?](#what-is-rmarkdown-and-how-can-i-install-it-optional)
-=======
->>>>>>> pkgdown
 
 # EvolVEd
 
@@ -144,563 +130,3 @@ covers the following topics, organized from basic to advanced:
 - Estimating diversification rates under pure birth and birth-death
   models
 - Factors influencing speciation and extinction
-<<<<<<< HEAD
-
-## Community guidelines and suggesting changes
-
-Questions and support requests should be sent to Matheus Januario’s
-email (januarioml.eco \[at\] gmail.com). Bug & issue reports,
-suggestions, improvements, or code additions should be pushed through
-the same GitHub repository. Emailing Matheus Januario is also ok.
-
-# If you never used R before…
-
-If you’re new to R, this “pre-vignette” tutorial serves as an
-introduction to using `evolved`. It was largely inspired by the first
-tutorials in EEB 485 and EEB 391/516 (Population and Community Ecology &
-Evolutionary processes and Macroevolution), but it is extended and
-modified to work in our context. This document is meant to serve as a
-first contact with the tools that will be of major importance in our
-vignettes with `evovled`. Nothing related to this tutorial is required
-to use the R package, but we will be using all the tools explained here,
-so please make sure you have completed it before then (or to be genrlly
-familiar with R).
-
-This tutorial will be somewhat long, and if this is your first time
-using R, it will also be a bit confusing. This is because a reasonable
-part of what you will see here is pretty abstract. We ask you to trust
-the learning process and keep reading until the end. We highly encourage
-you to reach out to other peers more knowledgeable in R, if needed. We
-know that different folks come with very different levels of comfort
-with R – which means you will learn a lot with your colleagues, however
-comfortable you may currently be with R.
-
-## What is R and how can I install it?
-
-R is (1) an open source software, (2) a programming language, and (3) a
-programming environment. The first one means it is a computer program
-you can freely download, use, or modify to your own needs. The second
-means it is a language that allows humans and machines to communicate.
-The third means it gives you everything you need to create your own
-programs under a specific language. Scientists, especially in the
-biological sciences, widely use R to run simulations and analyze their
-datasets, so we will do this too.
-
-If you’re using your own computer and have not yet installed R, you can
-download and install it for free by pasting this in your web browser:
-
-<http://cran.us.r-project.org/>
-
-After you download it, install it as you would do with any other program
-– the installer program will direct you though the major steps. When you
-open R, you should see something close to Figure 1.
-
-Please download and install R before progressing to install RStudio.
-
-<div class="figure" style="text-align: center">
-
-<img src="README_images/R_screenshot.png" alt="Example of recently-opened R on an Apple computer. The interface looks slightly different on a Windows computer, but the main features are the same. Note that depending on the details, you may or may not see the same warnings and errors as the screenshot." width="100%" />
-<p class="caption">
-Example of recently-opened R on an Apple computer. The interface looks
-slightly different on a Windows computer, but the main features are the
[…]
