### 2025-11-28 a6bad52
addgin comm guidelines 2 readme
