+++ README.Rmd
@@ -112,3 +112,8 @@ With `"vignette_name"` being one of the names below. Each vignette covers the fo
 #### 9. Phylogenies and Birth-Death Models (name: `birthdeath_phylogenies`)
 * Estimating diversification rates under pure birth and birth-death models
 * Factors influencing speciation and extinction
+
+## Community guidelines and suggesting changes
+
+Questions and support requests should be sent to Matheus Januario's email (januarioml.eco [at] gmail.com). 
+Bug & issue reports, suggestions, improvements, or code additions should be pushed through the same GitHub repository. Emailing Matheus Januario is also ok.
