### 2025-12-01 372fd1e
Update paper.md
adding lweb link to pkgdown page
