+++ README.Rmd
@@ -19,7 +19,7 @@ knitr::opts_chunk$set(echo = TRUE)
 
 # EvolVEd
 
-`evolved` (EVOLutionary Virtual EDucation) is an open-source R package designed for graduate or advanced undergraduate courses in evolutionary biology. It emphasizes tools for inquiry-based learning, where students engage in scientific practices to actively build knowledge (Pedaste et al., 2015). The package includes `vignettes` (tutorials) to facilitate classroom investigations. However, educators are encouraged to develop their own content modules depending on class context and/or learning objectives. Most of evolved‘s core functions are oriented towards either ($i$) developing student intuition for classic models in evolutionary biology using simulations, or ($ii$) analyzing (parameter estimation, hypothesis testing, visualization) datasets associated with key questions in evolutionary biology.
+`evolved` (EVOLutionary Virtual EDucation) is an open-source R package designed for graduate or advanced undergraduate courses in evolutionary biology. It emphasizes tools for inquiry-based learning, where students engage in scientific practices to actively build knowledge (Pedaste et al., 2015). The package includes `vignettes` (tutorials) to facilitate classroom investigations. They can be accessed online at [https://mjanuario.github.io/evolved/](https://mjanuario.github.io/evolved/) However, educators are encouraged to develop their own content modules depending on class context and/or learning objectives. Most of evolved‘s core functions are oriented towards either ($i$) developing student intuition for classic models in evolutionary biology using simulations, or ($ii$) analyzing (parameter estimation, hypothesis testing, visualization) datasets associated with key questions in evolutionary biology.
 
 ## Installing EvolVEd
 
@@ -63,19 +63,19 @@ vignette("vignette_name", package = "evolved")
 
 With `"vignette_name"` being one of the names below. Each vignette covers the following topics, organized from basic to advanced:
 
-#### 1. If you never used R before... (name: `install_r`)
+#### 1. If you never used R before... (name: [`install_r`](https://mjanuario.github.io/evolved/articles/install_r.html))
 * Introduces R and guides complete beginners through installation.
 * Explains RStudio’s interface and recommended configurations.
 * Teaches essential workflow: directories, saving, knitting, sourcing, loading data.
 * Troubleshoots common R/RMarkdown issues for new users.
 * Presents RMarkdown basics and why it’s useful for scientific work.
 
-#### 2. Intro to R (name: `intro_r`)
+#### 2. Intro to R (name: [`intro_r`](https://mjanuario.github.io/evolved/articles/intro_r.html))
 * Basic R syntax and coding (objects, vector calculations, etc.)
 * Plotting and annotation functions
 * Overview of key object classes used in the vignettes
 
-#### 3. Introduction to Population Genetics (name: `popgen_intro`)
+#### 3. Introduction to Population Genetics (name: [`popgen_intro`](https://mjanuario.github.io/evolved/articles/popgen_intro.html))
 * Simple mathematical notation
 * Probability of independent events and random number generation
 * Malthusian growth and Mendelian genetics
@@ -83,33 +83,33 @@ With `"vignette_name"` being one of the names below. Each vignette covers the fo
 * Heterozygosity, deleterious alleles, and mutation
 * Genetics at multiple loci and DNA variation
 
-#### 4. Genetic Drift (name: `popgen_drift`)
+#### 4. Genetic Drift (name: [`popgen_drift`](https://mjanuario.github.io/evolved/articles/popgen_drift.html))
 * Intuition and qualitative expectations of genetic drift
 * Variability in outcomes and heterozygosity decay
[…]
