+++ pkgdown.yml
@@ -14,7 +14,7 @@ articles:
   popgen_selection: popgen_selection.html
   population_genetics: population_genetics.html
   rocks_and_clocks: rocks_and_clocks.html
-last_built: 2025-11-28T20:45Z
+last_built: 2025-12-01T16:57Z
 urls:
   reference: https://mjanuario.github.io/evolved/reference
   article: https://mjanuario.github.io/evolved/articles
