+++ pkgdown.yml
@@ -14,7 +14,7 @@ articles:
   popgen_selection: popgen_selection.html
   population_genetics: population_genetics.html
   rocks_and_clocks: rocks_and_clocks.html
-last_built: 2025-12-01T17:23Z
+last_built: 2025-12-01T17:28Z
 urls:
   reference: https://mjanuario.github.io/evolved/reference
   article: https://mjanuario.github.io/evolved/articles
