+++ CONTRIBUTING.md.R
@@ -1,0 +1,20 @@
+## How to Contribute
+
+* Vignettes can be acessed at [https://mjanuario.github.io/evolved/](https://mjanuario.github.io/evolved/), or by downloading the package from CRAN or github (see instructions in the website).
+* If you want to contribute with functions or tutorials, please reach out to Matheus Januario (januarioml.eco [at] gmail.com), and we can work with you to add new material and data to the site
+* Alternatively, your contributions or new lessons are welcome as Pull Requests.
+
+## Reporting Issues or Problems using github
+
+* To report issues, bugs, or problems with either the material, the functions, or the wbesite please [open an issue on our github page](https://github.com/mjanuario/evolved/issues/new) and we'll be happy to address it.
+* ALternatively, send an email to Matheus Januario (januarioml.eco [at] gmail.com).
+
+
+## Support
+
+* Send an email to Matheus Januario (januarioml.eco [at] gmail.com).
+* [Open an issue](https://github.com/mjanuario/evolved/issues/new) if there’s anything we can do to help!
+
+## License
+
+See our [License](https://github.com/mjanuario/evolved/blob/main/LICENSE.md) for information on copyright and ownership. Questions can also be sent to to Matheus Januario (januarioml.eco [at] gmail.com).
