### 2025-12-01 14fcf15
adding contributing to readme
