### 2025-12-01 a3c2735
adding badge to your README.Rmd
