### 2025-12-01 02a5e7e
Update paper.md
added AV's orcid
