### 2025-12-01 851a667
Update paper.md
more links to website
