### 2025-12-01 ed76790
Update paper.md
typo fix
