### 2025-12-01 #300 @mjanuario
Dear @ethanwhite and @fmichonneau,

Apologies for the delay in our response. We needed some time to familiarize ourselves with a few aspects, but we have now addressed, to the best of our knowledge, all the points raised and implemented several additional improvements:

1. The pkgdown website is now available at https://mjanuario.github.io/evolved/ , providing a comprehensive online reference for the package.

2. A GitHub Actions workflow has been added to test vignette builds and the pkgdown site, ensuring that any issues are detected prior to merging.

3. The CONTRIBUTING.md file is available at https://github.com/mjanuario/evolved/blob/main/CONTRIBUTING.md and is linked from both the README and the website.

4. Community guidelines have been added in multiple locations: at the end of the README (with a link at the top), in a dedicated “Community Guidelines” section on the GitHub page, in the CONTRIBUTING.md file, and in the manuscript itself.

5. Unnecessary files such as DS.STORE have been removed from all branches in the repository.

6. We have released version 1.0.0 of the package on GitHub, and this version is now referenced in the manuscript.

7. All of the above updates are incorporated in the PDF version of the manuscript.

We would greatly appreciate any feedback if something was overlooked or if any of our implementations were incorrect.

Thank you so much for your guidance and feedback throughout this process and hope these updates fully address the previous comments.

Best regards,
Matheus Januario (on behalf of all authors)
