### 2025-12-05 #300 @ethanwhite
Thanks for your excellent work on this @mjanuario. It looks like your new GitHub Actions workflow (thanks so much for adding it!) is currently failing.

Take a look and see if you can debug the issue and let us know if you get stuck and could use another set of eyeballs on it.
