+++ reference/OneGenHWSim.md
@@ -1,0 +1,284 @@
+# Simulating one generation of genotypes under Hardy-Weinberg equilibrium
+
+`OneGenHWSim` creates `n.sim` simulations of one generation of genotypes
+under Hardy-Weinberg equilibrium for a bi-allelic loci.
+
+## Usage
+
+``` r
+OneGenHWSim(n.ind = 50, p = 0.5, n.sim = 100)
+```
+
+## Arguments
+
+- n.ind:
+
+  Integer indicating the census size of the simulated populations. If
+  decimals are inserted, they will be rounded.
+
+- p:
+
+  Numerical between zero and one that indicates A1's allele frequency.
+  A2's allele frequency is assumed to be `1-p`.
+
+- n.sim:
+
+  Number of simulations to be made. If decimals are inserted, they will
+  be rounded.
+
+## Value
+
+A `data.frame` containing the number of individuals for each genotype.
+
+## References
+
+Hardy, G. H. (1908). Mendelian proportions in a mixed population.
+Science, 28, 49–50.
+
+Weinberg, W. (1908). Uber den Nachweis der Vererbung beim Menschen.
+Jahreshefte des Vereins fur vaterlandische Naturkunde in Wurttemberg,
+Stuttgart 64:369–382. \[On the demonstration of inheritance in humans\].
+Translation by R. A. Jameson printed in D. L. Jameson (Ed.), (1977).
+Benchmark papers in genetics, Volume 8: Evolutionary genetics (pp.
+115–125). Stroudsburg, PA: Dowden, Hutchinson & Ross.
+
+Mayo, O. (2008). A century of Hardy–Weinberg equilibrium. Twin Research
+and Human Genetics, 11(3), 249-256.
+
+## See also
+
+Other microevolution:
+[`NatSelSim()`](https://mjanuario.github.io/evolved/reference/NatSelSim.md),
+[`WFDriftSim()`](https://mjanuario.github.io/evolved/reference/WFDriftSim.md),
+[`plotNatSel()`](https://mjanuario.github.io/evolved/reference/plotNatSel.md),
+[`plotWFDrift()`](https://mjanuario.github.io/evolved/reference/plotWFDrift.md)
+
+## Author
+
+Matheus Januario, Dan Rabosky, Jennifer Auler
+
+## Examples
+
+``` r
+#using the default values (n.ind = 50, p = 0.5, n.sim = 100):
+OneGenHWSim()
+#>         A1A1 A1A2 A2A2
+#> sim_1     13   24   13
+#> sim_2     13   24   13
+#> sim_3     10   30   10
+#> sim_4     12   26   12
+#> sim_5     12   26   12
+#> sim_6     13   24   13
+#> sim_7     14   22   14
+#> sim_8     13   24   13
+#> sim_9     12   26   12
+#> sim_10     9   32    9
+#> sim_11    13   24   13
+#> sim_12    14   22   14
+#> sim_13    11   28   11
+#> sim_14    15   20   15
+#> sim_15    10   30   10
+#> sim_16    10   30   10
+#> sim_17    13   24   13
+#> sim_18    14   22   14
+#> sim_19    12   26   12
+#> sim_20     9   32    9
+#> sim_21     9   32    9
+#> sim_22    12   26   12
+#> sim_23    10   30   10
+#> sim_24    10   30   10
+#> sim_25    12   26   12
+#> sim_26    11   28   11
+#> sim_27    12   26   12
+#> sim_28    10   30   10
+#> sim_29    11   28   11
+#> sim_30    13   24   13
+#> sim_31    11   28   11
[…]
