+++ reference/WFDriftSim.md
@@ -1,154 +1,0 @@
-# Simulating generations of genetic drift in a Wright–Fisher (WF) population
-
-`WFDriftSim` simulates genetic drift of diploid Wright–Fisher
-populations with a given effective population size through a certain
-number of generations.
-
-## Usage
-
-``` r
-WFDriftSim(
-  Ne,
-  n.gen,
-  p0 = 0.5,
-  n.sim = 1,
-  plot.type = "animate",
-  print.data = FALSE,
-  knitr = FALSE
-)
-```
-
-## Arguments
-
-- Ne:
-
-  Number giving the effective population size of the population
-
-- n.gen:
-
-  Number of generations to be simulated.
-
-- p0:
-
-  Initial frequency of a given allele. As the simulated organism is
-  diploid, the other alleles frequency will be `1-(p0)`. Default value
-  is `0.5`.
-
-- n.sim:
-
-  Number of simulations to be made. If decimals are inserted, they will
-  be rounded. Default value is `1`.
-
-- plot.type:
-
-  Character indicating if simulations should be plotted as colored
-  lines. Each color represents a different population. If
-  `plot.type = "animate"` (default value) it animates each generation
-  individually. If `plot.type = "static"` it plots all lines rapidly. If
-  `plot.type = "none"` nothing is plotted.
-
-- print.data:
-
-  Logical indicating whether all simulation results should be returned
-  as a `data.frame`. Default value is `FALSE`.
-
-- knitr:
-
-  Logical indicating if plot is intended to show up in RMarkdown files
-  made by the `Knitr` R package.
-
-## Value
-
-If `plot.type = "static"` or `"animate"`, plots the timeseries of all
-simulations, with each line+color referring to a different simulation.
-Note that if many simulations (generally more than 20) are simulated,
-colors might be cycled and different simulation will have the same
-color. If `print.data = TRUE`, returns a `data.frame` with the
-simulation results.
-
-## Details
-
-The effective population size (`Ne`) is strongly connected with the rate
-of genetic drift (for details, see Waples, 2022).
-
-## References
-
-Fisher RA (1922) On the dominance ratio. Proc. R. Soc. Edinb 42:321–341
-
-Kimura M (1955) Solution of a process of random genetic drift with a
-continuous model. PNAS–USA 41(3):144–150
-
-Tran, T. D., Hofrichter, J., & Jost, J. (2013). An introduction to the
-mathematical structure of the Wright–Fisher model of population
-genetics. Theory in Biosciences, 132(2), 73-82. \[good for the
-historical review, math can be challenging\]
-
-Waples, R. S. (2022). What is Ne, anyway?. Journal of Heredity.
-
-Wright S (1931) Evolution in Mendelian populations. Genetics 16:97–159
-
-## See also
-
-Other microevolution:
-[`NatSelSim()`](https://mjanuario.github.io/evolved/reference/NatSelSim.md),
-[`OneGenHWSim()`](https://mjanuario.github.io/evolved/reference/OneGenHWSim.md),
[…]
