+++ reference/birds_spp.md
@@ -1,26 +1,0 @@
-# Birds species list
-
-Birds species list
-
-## Usage
-
-``` r
-birds_spp
-```
-
-## Format
-
-A `vector` containing the names of almost all (i.e., 9993) extant bird
-species, following Jetz et al. (2012) taxonomy. This dataset is part of
-the package and is licensed under the Creative Commons Attribution 4.0
-International License (CC BY 4.0).
-
-## Source
-
-Actual file downloaded from <https://vertlife.org/data/>
-
-## References
-
-Jetz, W., Thomas, G. H., Joy, J. B., Hartmann, K., & Mooers, A. O.
-(2012). The global diversity of birds in space and time. Nature,
-491(7424), 444-448.
