+++ articles/rocks_and_clocks.md
@@ -1,294 +1,0 @@
-# Clocks and Rocks
-
-``` r
-library(evolved)
-
-# Let's also store our par() configs so 
-# we can restore them whenever we change it in this tutorial
-oldpar <- par(no.readonly = TRUE)  
-```
-
-## The fossil record:
-
-Here we will use data on fossilized organism, which we call “fossil
-occurrences”. the data was originally obtained from the Paleobiology
-database, a free-to-use resource that is the standard repository for
-such data and which has been used by hundreds or thousands of scientific
-research articles to study the history and dynamics of biodiversity in
-Deep Time.
-
-``` r
-data("dinos_fossil")
-head(dinos_fossil)
-```
-
-    ##     phylum        class              order              family
-    ## 1 Chordata Ornithischia NO_ORDER_SPECIFIED    Chaoyangsauridae
-    ## 2 Chordata   Saurischia       Avetheropoda NO_FAMILY_SPECIFIED
-    ## 3 Chordata   Saurischia       Avetheropoda NO_FAMILY_SPECIFIED
-    ## 4 Chordata   Saurischia       Avetheropoda     Tyrannosauridae
-    ## 5 Chordata   Saurischia       Avetheropoda     Tyrannosauridae
-    ## 6 Chordata Ornithischia NO_ORDER_SPECIFIED        Ceratopsidae
-    ##               genus                   species early_interval late_interval
-    ## 1    Chaoyangsaurus     Chaoyangsaurus youngi Late Tithonian   Valanginian
-    ## 2 Protarchaeopteryx Protarchaeopteryx robusta Late Barremian  Early Aptian
-    ## 3       Caudipteryx          Caudipteryx zoui Late Barremian  Early Aptian
-    ## 4       Gorgosaurus      Gorgosaurus libratus Late Campanian              
-    ## 5       Gorgosaurus      Gorgosaurus libratus Late Campanian              
-    ## 6      Centrosaurus      Centrosaurus apertus Late Campanian              
-    ##   max_ma min_ma midpoint       lng      lat
-    ## 1  150.8 132.90   141.85  123.9667 42.93330
-    ## 2  130.0 122.46   126.23  120.7333 41.80000
-    ## 3  130.0 122.46   126.23  120.7333 41.80000
-    ## 4   83.5  70.60    77.05 -111.5287 50.74073
-    ## 5   83.5  70.60    77.05 -111.5493 50.73701
-    ## 6   83.5  70.60    77.05 -111.5289 50.73730
-
-Students can use functions in the package to calculate the diversity of
-a certain taxonomic rank, and plot it through time.
-
-To make it more fun, we will compare tow different taxonomic levels, and
-will plot them in a relative, log scale:
-
-``` r
-spDTT = calcFossilDivTT(dinos_fossil, tax.lvl = "species")
-genusDTT = calcFossilDivTT(dinos_fossil, tax.lvl = "genus")
-famDTT = calcFossilDivTT(dinos_fossil, tax.lvl = "family")
-
-# And to allow comparisons, we will use relative richness:
-plot(x=genusDTT$age, xlim = rev(range(genusDTT$age)),
-     y=log(genusDTT$div)-log(max(genusDTT$div)), 
-     xlab="Time (Million years ago)",
-     ylab="Log relative diversity",
-     type="l", col="blue", ylim=c(-7,0))
-
-lines(x=famDTT$age,
-     y=log(famDTT$div)-log(max(famDTT$div)), 
-     col="red")
-
-lines(x=spDTT$age,
-     y=log(spDTT$div)-log(max(spDTT$div)), 
[…]
