### 2025-12-17 4b40256
1st attemp to fix plot3D for MAcOS without x11
