+++ articles/birthdeath_phylogenies.md
@@ -1,0 +1,644 @@
+# Birth-death models and molecular phylogenies
+
+## Learning objectives
+
+1.  Estimating diversification rates under pure birth
+2.  Estimating diversification rates under birth-death
+3.  What affects speciation and extinction?
+
+## Introduction
+
+First, we will load our required packages:
+
+``` r
+library(LSD)
+library(ape)
+library(BAMMtools)
+library(evolved)
+
+# Let's also store our par() configs so 
+# we can restore them whenever we change it in this tutorial
+oldpar <- par(no.readonly = TRUE)  
+```
+
+## Simulating phylogenetic trees
+
+Molecular phylogenies are built from DNA sequences and therefore only
+include extant species. In this vignette, we will simulate and analyze
+“ideal” versions of these molecular phylogenies. We know from previous
+vignettes that using molecular phylogenies directly (“naively”) can lead
+to spurious conclusions, as these phylogenies are missing a lot of
+information. Despite this limitation, there is still a lot we can learn
+about speciation and extinction from molecular phylogenies. This is what
+we will do today.
+
+As we dive into the realm of molecular phylogenies, keep in mind that
+most groups of organisms do not have good fossil records. In such cases,
+phylogenetic trees of extant species are often the only information we
+will ever have about rates of speciation and/or extinction in these
+groups.
+
+We will be simulating molecular phylogenies with a function from our
+package `evolved`, so let’s first check the arguments of
+[`simulateTree()`](https://mjanuario.github.io/evolved/reference/simulateTree.md).
+
+``` r
+args(simulateTree)
+```
+
+    ## function (pars, max.taxa = Inf, max.t, min.taxa = 2, include.extinct = FALSE) 
+    ## NULL
+
+`pars =` A vector containing the speciation and extinction rates,
+written as `c(S, E)`.
+
+`max.taxa =` Maximum number of taxa to include in the tree.
+
+`min.taxa =` Minimum number of taxa to include in the tree.
+
+`max.t =` Maximum time of simulation.
+
+`include.extinct =` A logical (TRUE or FALSE) describing if the
+phylogeny should include lineages that went extinct before the
+simulation ended.
+
+This function carries out a simulation of diversification, resulting in
+a phylogenetic tree that traces the relationships between the taxa that
+speciated and went extinct during the simulation. Speciation is
+simulated as one lineage “giving birth” to another, bifurcating into two
+new species. Extinction is of course the complete disappearance or
+“death” of a lineage. Whether these events occur or do not occur is
+dependent on a random number draw. Also, the simulation keeps track of
+how many time steps have passed since its initiation.
+
+The simulation ends when either (a) the amount of elapsed time surpasses
+the maximum time for the simulation, given in the parameter `max.t`, or
+(b) the number of lineages existing (“alive”) at the current time step
+surpasses the maximum number of taxa, given in the parameter `max.taxa`.
+These two cases are the “stopping criteria” for the simulation, and when
+either occurs, the simulation terminates and outputs a phylogenetic
+tree.
+
+Now that you understand what is going on “under the hood” in this
+simulation, let’s try to run one. Initiate a simulation as follows:
+
+``` r
+# Set parameters for simulation
+S <- 1
+E <- 0
+
+# Set seed so we all get same results
+set.seed(1)
+
+# First, generate tree with no extinction and 6 tips
[…]
