+++ reference/mammals_spp.md
@@ -1,0 +1,26 @@
+# Mammals species list
+
+Mammals species list
+
+## Usage
+
+``` r
+mammals_spp
+```
+
+## Format
+
+A `vector` containing the names of almost all (i.e., 4099) extant mammal
+species, following Upham et al (2019) taxonomy. This dataset is part of
+the package and is licensed under the Creative Commons Attribution 4.0
+International License (CC BY 4.0).
+
+## Source
+
+Actual file downloaded from <https://vertlife.org/data/>
+
+## References
+
+Upham, N. S., Esselstyn, J. A., & Jetz, W. (2019). Inferring the mammal
+tree: species-level sets of phylogenies for questions in ecology,
+evolution, and conservation. PLoS biology, 17(12), e3000494.
