### 2025-12-17 f4961fc
removed plot3D from tutorial
