### 2025-12-17 6317d1c
removed plot3D from tutorial
