+++ articles/birthdeath_phylogenies.md
@@ -230,9 +230,9 @@ matrix(c(estimateSpeciation(tree1), estimateSpeciation(tree2), estimateSpeciatio
 ```
 
     ##      Estimated True
-    ## [1,]  0.533297  0.5
-    ## [2,]  0.774391  1.0
-    ## [3,]  4.445254  5.0
+    ## [1,] 0.3768694  0.5
+    ## [2,] 1.0826606  1.0
+    ## [3,] 4.4620947  5.0
 
 Now that you have a sense for the accuracy of our estimator, let’s
 simulate 100 phylogenetic trees with variable speciation rates and apply
@@ -356,7 +356,7 @@ fitCRBD(tree)
 ```
 
     ##        S        E 
-    ## 1.318936 1.300098
+    ## 1.294217 1.277742
 
 Let’s repeat our simulation procedure, but this time we will fit a model
 where extinction can be nonzero. We won’t display the plots here, but
