### 2026-01-06 #300 @ethanwhite
Thanks @mjanuario! Everything looks good now!

@fmichonneau can you check to confirm that @mjanuario's changes have addressed your comments. If so, please complete the remaining items on your checklist and let me know you're happy with the software/paper. If not, please provide additional feedback for @mjanuario.
