### 2026-01-09 #300 @fmichonneau
Thank you @mjanuario for your work on this. A took a quick look at the GitHub repo and I appreciate the effort you put into adding the pkgdown website as well as the contributing guidelines.

@ethanwhite I'm happy with the software/paper being published. Thanks!
