### 2026-01-09 #300 @ethanwhite
@mjanuario - we just have a few more things to do to get everything wrapped up. I've generated the post-review checklist. Look through the items in the author tasks section (I'm not sure if you can check them off or not) and make sure they are ready to go. Since you've already made a release the biggest thing is setting up the archive. I find the easiest way to do this is through the [GitHub - Zenodo integration](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) in which case you will need to setup the integration first and then make a new release in GitHub (this can just be a patch release to trigger the archive). Alternatively it is fine to manually upload a copy of the repository to Zenodo. The author line and title on Zenodo should match the paper and the license and release tag on Zenodo should match those on GitHub. One way to do this automatically with the integration is to add a [.zenodo.json file](https://help.zenodo.org/docs/github/describe-software/zenodo-json/) to the GitHub repository prior to the release. If you have any questions about this step let me know.

I'll be performing other editorial checks and let you know if there is anything else you need to do.
