### 2026-01-09 #300 @ethanwhite
After a final read through the paper I have just a couple of editorial suggestions:

- The word "Evolutionary" in the title should not be capitalized
- Suggest removing the `-` from the citation at the end of the first paragraph for consistency with similar parenthetical citations
- When `R` is included in the title of papers in the References section it should be capitalized (there are ~6 cases where it is not).
