### 2026-01-09 e334371
Update paper.md
article adjusts
