### 2026-01-09 886aad3
Update paper.bib
fixing capitalization in bibtex
