+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{BanchiBell:2008,
-  title={The many levels of inquiry.},
+  title={The many levels of inquiry},
   author={Banchi, Heather and Bell, Randy},
   journal={Science \& Children},
   volume={46},
@@ -21,7 +21,7 @@
 }
 
 @article{Fitzjohn:2012,
-  title={Diversitree: comparative phylogenetic analyses of diversification in R},
+  title={Diversitree: comparative phylogenetic analyses of diversification in {R}},
   author={FitzJohn, Richard G},
   journal={Methods in Ecology and Evolution},
   volume={3},
@@ -64,7 +64,7 @@
 }
 
 @article{Paradis:2019,
-  title={ape 5.0: an environment for modern phylogenetics and evolutionary analyses in R},
+  title={ape 5.0: an environment for modern phylogenetics and evolutionary analyses in {R}},
   author={Paradis, Emmanuel and Schliep, Klaus},
   journal={Bioinformatics},
   volume={35},
@@ -96,7 +96,7 @@
 }
 
 @article{Rabosky:2014,
-  title={BAMM tools: an R package for the analysis of evolutionary dynamics on phylogenetic trees},
+  title={BAMM tools: an {R} package for the analysis of evolutionary dynamics on phylogenetic trees},
   author={Rabosky, Daniel L and Grundler, Michael and Anderson, Carlos and Title, Pascal and Shi, Jeff J and Brown, Joseph W and Huang, Huateng and Larson, Joanna G},
   journal={Methods in Ecology and Evolution},
   volume={5},
@@ -120,7 +120,7 @@
 }
 
 @article{Revell:2024,
-  title={phytools 2.0: an updated R ecosystem for phylogenetic comparative methods (and other things)},
+  title={phytools 2.0: an updated {R} ecosystem for phylogenetic comparative methods (and other things)},
   author={Revell, Liam J},
   journal={PeerJ},
   volume={12},
