### 2026-01-09 8364746
Update paper.bib
more adj in bibtex
