+++ paper/paper.bib
@@ -88,7 +88,7 @@
 
 @article{R:2024,
   title={{R: A language and environment for statistical computing}},
-  author={R Core Team},
+  author={{R Core Team}},
   journal={Foundation for Statistical Computing, Vienna, Austria},
   year={2024},
   url = {https://www.R-project.org/},
