### 2026-01-09 aa55ec8
Update paper.md
added funding to acknol section
