### 2026-01-09 dd12063
Create CITATION.cff
added first version
