+++ CITATION.cff
@@ -1,0 +1,55 @@
+cff-version: 1.2.0
+message: "If you use this software, please cite it as below."
+
+title: "evolved: Open Software for Teaching Evolutionary Biology at Multiple Scales Through Virtual Inquiries"
+
+type: software
+
+authors:
+  - family-names: Januario
+    given-names: Matheus
+    orcid: https://orcid.org/0000-0002-6480-7095
+    affiliation: Museum of Zoology and Department of Ecology and Evolutionary Biology, University of Michigan, Ann Arbor, MI, USA
+
+  - family-names: Auler
+    given-names: Jennifer P.
+    orcid: https://orcid.org/0000-0001-7576-9058
+    affiliation: Programa de Pós-Graduação em Ecologia, Universidade de São Paulo, São Paulo, SP, Brazil
+
+  - family-names: Viol
+    given-names: Andressa L.
+    orcid: https://orcid.org/0009-0004-2558-8389
+    affiliation: Museum of Zoology and Department of Ecology and Evolutionary Biology, University of Michigan, Ann Arbor, MI, USA
+
+  - family-names: Rabosky
+    given-names: Daniel L.
+    orcid: https://orcid.org/0000-0002-7499-8251
+    affiliation: Museum of Zoology and Department of Ecology and Evolutionary Biology, University of Michigan, Ann Arbor, MI, USA
+
+version: 1.0.1
+
+license: GPL-3.0
+
+keywords:
+  - evolutionary biology
+  - phylogenetics
+  - population genetics
+  - paleobiology
+  - macroevolution
+  - inquiry-based learning
+  - science education
+  - simulation
+  - R
+
+repository-code: https://github.com/mjanuario/evolved
+url: https://cran.r-project.org/package=evolved
+
+date-released: 2024-12-19
+
+funding:
+  - name: National Science Foundation
+    identifier: https://www.nsf.gov
+    grant-number: 1754398
+
+  - name: Rackham Predoctoral Fellowship
+    identifier: https://rackham.umich.edu
