+++ articles/birthdeath_phylogenies.md
@@ -230,9 +230,9 @@ matrix(c(estimateSpeciation(tree1), estimateSpeciation(tree2), estimateSpeciatio
 ```
 
     ##      Estimated True
-    ## [1,] 0.4378347  0.5
-    ## [2,] 1.0232958  1.0
-    ## [3,] 4.5317943  5.0
+    ## [1,] 0.4741897  0.5
+    ## [2,] 1.1573038  1.0
+    ## [3,] 4.4019975  5.0
 
 Now that you have a sense for the accuracy of our estimator, let’s
 simulate 100 phylogenetic trees with variable speciation rates and apply
@@ -356,7 +356,7 @@ fitCRBD(tree)
 ```
 
     ##        S        E 
-    ## 1.714802 1.681248
+    ## 1.440485 1.497985
 
 Let’s repeat our simulation procedure, but this time we will fit a model
 where extinction can be nonzero. We won’t display the plots here, but
