### 2026-01-09 69134f1
Update CITATION.cff
minimalist version for zenodo
