### 2026-01-09 70124d1
Update CITATION.cff
minimalist version for zenodo
