### 2026-01-09 2ed4bb2
Update CITATION.cff
for zenodo
