### 2026-01-09 2afaa96
Update README.Rmd
adding zenodo's DOI
