+++ articles/birthdeath_phylogenies.md
@@ -230,9 +230,9 @@ matrix(c(estimateSpeciation(tree1), estimateSpeciation(tree2), estimateSpeciatio
 ```
 
     ##      Estimated True
-    ## [1,] 0.6805893  0.5
-    ## [2,] 0.9717245  1.0
-    ## [3,] 5.8880453  5.0
+    ## [1,] 0.4534433  0.5
+    ## [2,] 0.8735446  1.0
+    ## [3,] 4.6884374  5.0
 
 Now that you have a sense for the accuracy of our estimator, let’s
 simulate 100 phylogenetic trees with variable speciation rates and apply
@@ -356,7 +356,7 @@ fitCRBD(tree)
 ```
 
     ##        S        E 
-    ## 1.429074 1.445892
+    ## 1.533514 1.152733
 
 Let’s repeat our simulation procedure, but this time we will fit a model
 where extinction can be nonzero. We won’t display the plots here, but
