+++ articles/birthdeath_phylogenies.md
@@ -230,9 +230,9 @@ matrix(c(estimateSpeciation(tree1), estimateSpeciation(tree2), estimateSpeciatio
 ```
 
     ##      Estimated True
-    ## [1,] 0.4182124  0.5
-    ## [2,] 0.9805864  1.0
-    ## [3,] 4.9843589  5.0
+    ## [1,] 0.5528097  0.5
+    ## [2,] 0.7578983  1.0
+    ## [3,] 4.8133745  5.0
 
 Now that you have a sense for the accuracy of our estimator, let’s
 simulate 100 phylogenetic trees with variable speciation rates and apply
@@ -356,7 +356,7 @@ fitCRBD(tree)
 ```
 
     ##        S        E 
-    ## 1.264468 1.243084
+    ## 1.798701 1.867530
 
 Let’s repeat our simulation procedure, but this time we will fit a model
 where extinction can be nonzero. We won’t display the plots here, but
