### 2026-01-10 #300 @mjanuario
Many thanks @ethanwhite !

All suggestions have been accepted and implemented but I only found 4 instances of low-caps R in the refs.

DOI of zenodo archive is: [10.5281/zenodo.18203397](https://doi.org/10.5281/zenodo.18203397)

Licenses & ORCIDS do match.
