### 2026-01-12 #300 @ethanwhite
Thanks @mjanuario. Now that the archive is up there are a couple of small fixes to make sure that it matches the paper and repository:

- The version number on GitHub is `v1.0.22` and the version number on Zenodo is `1.0.1`. Both the last number and the presence or absence of a `v` at the start of the version number should match.
- The titles should also match precisely. In the paper the title is `evolved: open software for teaching evolutionary biology through virtual inquiries` on Zenodo it is `evolved: Open Software for Teaching Evolutionary Biology`

Once these are fixed I should be ready to wrap this up and move it on to the Editor in Chief.
