### 2026-01-12 bb918cf
Update CITATION.cff
quick caps title fix for zenodo
