+++ CITATION.cff
@@ -1,7 +1,7 @@
 cff-version: 1.2.0
 message: "If you use this software, please cite it as below."
 
-title: "evolved: Open Software for Teaching Evolutionary Biology at Multiple Scales Through Virtual Inquiries"
+title: "evolved: open software for teaching evolutionary biology at multiple scales through virtual inquiries"
 
 type: software
 
