+++ articles/birthdeath_phylogenies.md
@@ -230,9 +230,9 @@ matrix(c(estimateSpeciation(tree1), estimateSpeciation(tree2), estimateSpeciatio
 ```
 
     ##      Estimated True
-    ## [1,] 0.4578884  0.5
-    ## [2,] 1.1708238  1.0
-    ## [3,] 5.1397747  5.0
+    ## [1,] 0.5298674  0.5
+    ## [2,] 0.9499086  1.0
+    ## [3,] 5.3631094  5.0
 
 Now that you have a sense for the accuracy of our estimator, let’s
 simulate 100 phylogenetic trees with variable speciation rates and apply
@@ -356,7 +356,7 @@ fitCRBD(tree)
 ```
 
     ##        S        E 
-    ## 1.530864 1.496703
+    ## 1.213361 1.188296
 
 Let’s repeat our simulation procedure, but this time we will fit a model
 where extinction can be nonzero. We won’t display the plots here, but
