### 2026-01-12 5e7bc8e
Update CITATION.cff
