### 2026-01-12 60000a9
Update CITATION.cff
