+++ CITATION.cff
@@ -30,8 +30,6 @@ license: GPL-3.0
 
 funding:
   - name: National Science Foundation
-    identifier: https://www.nsf.gov
     grant-number: 1754398
 
   - name: Rackham Predoctoral Fellowship
-    identifier: https://rackham.umich.edu
