+++ CITATION.cff
@@ -25,11 +25,3 @@ authors:
     given-names: Daniel L.
     orcid: https://orcid.org/0000-0002-7499-8251
     affiliation: Museum of Zoology and Department of Ecology and Evolutionary Biology, University of Michigan, Ann Arbor, MI, USA
-
-license: GPL-3.0
-
-funding:
-  - name: National Science Foundation
-    grant-number: 1754398
-
-  - name: Rackham Predoctoral Fellowship
