+++ articles/birthdeath_phylogenies.md
@@ -230,9 +230,9 @@ matrix(c(estimateSpeciation(tree1), estimateSpeciation(tree2), estimateSpeciatio
 ```
 
     ##      Estimated True
-    ## [1,]  0.585243  0.5
-    ## [2,]  1.033147  1.0
-    ## [3,]  4.259266  5.0
+    ## [1,] 0.4644171  0.5
+    ## [2,] 1.0379715  1.0
+    ## [3,] 5.9104048  5.0
 
 Now that you have a sense for the accuracy of our estimator, let’s
 simulate 100 phylogenetic trees with variable speciation rates and apply
@@ -356,7 +356,7 @@ fitCRBD(tree)
 ```
 
     ##        S        E 
-    ## 1.218722 1.153111
+    ## 1.387783 1.310753
 
 Let’s repeat our simulation procedure, but this time we will fit a model
 where extinction can be nonzero. We won’t display the plots here, but
