### 2026-01-12 ffcd989
Update CITATION.cff
