+++ CITATION.cff
@@ -10,18 +10,19 @@ authors:
     given-names: Matheus
     orcid: https://orcid.org/0000-0002-6480-7095
     affiliation: Museum of Zoology and Department of Ecology and Evolutionary Biology, University of Michigan, Ann Arbor, MI, USA
-
   - family-names: Auler
     given-names: Jennifer P.
     orcid: https://orcid.org/0000-0001-7576-9058
-    affiliation: Programa de Pós-Graduação em Ecologia, Universidade de São Paulo, São Paulo, SP, Brazil
-
+    affiliation: Programa de Pos-Graduacao em Ecologia, Universidade de Sao Paulo, Sao Paulo, SP, Brazil
   - family-names: Viol
     given-names: Andressa L.
     orcid: https://orcid.org/0009-0004-2558-8389
     affiliation: Museum of Zoology and Department of Ecology and Evolutionary Biology, University of Michigan, Ann Arbor, MI, USA
-
   - family-names: Rabosky
     given-names: Daniel L.
     orcid: https://orcid.org/0000-0002-7499-8251
     affiliation: Museum of Zoology and Department of Ecology and Evolutionary Biology, University of Michigan, Ann Arbor, MI, USA
+version: 1.0.23
+license: GPL-3.0
+repository-code: https://github.com/mjanuario/evolved
+url: https://mjanuario.github.io/evolved
