+++ articles/birthdeath_phylogenies.md
@@ -230,9 +230,9 @@ matrix(c(estimateSpeciation(tree1), estimateSpeciation(tree2), estimateSpeciatio
 ```
 
     ##      Estimated True
-    ## [1,] 0.4903825  0.5
-    ## [2,] 1.0023898  1.0
-    ## [3,] 6.3873873  5.0
+    ## [1,] 0.5597258  0.5
+    ## [2,] 0.8097964  1.0
+    ## [3,] 4.7882505  5.0
 
 Now that you have a sense for the accuracy of our estimator, let’s
 simulate 100 phylogenetic trees with variable speciation rates and apply
@@ -355,8 +355,8 @@ tree <- simulateTree(c(1, 0.95), max.taxa = 50, max.t = Inf)
 fitCRBD(tree)
 ```
 
-    ##        S        E 
-    ## 1.577785 1.515479
+    ##         S         E 
+    ## 1.0160477 0.6888858
 
 Let’s repeat our simulation procedure, but this time we will fit a model
 where extinction can be nonzero. We won’t display the plots here, but
