### 2026-01-12 5dba5f3
Update CITATION.cff
