+++ articles/birthdeath_phylogenies.md
@@ -230,9 +230,9 @@ matrix(c(estimateSpeciation(tree1), estimateSpeciation(tree2), estimateSpeciatio
 ```
 
     ##      Estimated True
-    ## [1,] 0.4569269  0.5
-    ## [2,] 1.0192027  1.0
-    ## [3,] 4.9019058  5.0
+    ## [1,] 0.3719898  0.5
+    ## [2,] 0.7688978  1.0
+    ## [3,] 5.0876131  5.0
 
 Now that you have a sense for the accuracy of our estimator, let’s
 simulate 100 phylogenetic trees with variable speciation rates and apply
@@ -356,7 +356,7 @@ fitCRBD(tree)
 ```
 
     ##        S        E 
-    ## 1.238593 1.246045
+    ## 1.977054 1.638573
 
 Let’s repeat our simulation procedure, but this time we will fit a model
 where extinction can be nonzero. We won’t display the plots here, but
