+++ articles/birthdeath_phylogenies.md
@@ -230,9 +230,9 @@ matrix(c(estimateSpeciation(tree1), estimateSpeciation(tree2), estimateSpeciatio
 ```
 
     ##      Estimated True
-    ## [1,] 0.3719898  0.5
-    ## [2,] 0.7688978  1.0
-    ## [3,] 5.0876131  5.0
+    ## [1,] 0.4891891  0.5
+    ## [2,] 0.6723873  1.0
+    ## [3,] 5.0687809  5.0
 
 Now that you have a sense for the accuracy of our estimator, let’s
 simulate 100 phylogenetic trees with variable speciation rates and apply
@@ -356,7 +356,7 @@ fitCRBD(tree)
 ```
 
     ##        S        E 
-    ## 1.977054 1.638573
+    ## 1.686256 1.597895
 
 Let’s repeat our simulation procedure, but this time we will fit a model
 where extinction can be nonzero. We won’t display the plots here, but
