### 2026-01-12 6f18302
Update CITATION.cff
