### 2026-01-24 e87d0f1
Update CITATION.cff
