+++ articles/birthdeath_phylogenies.md
@@ -230,9 +230,9 @@ matrix(c(estimateSpeciation(tree1), estimateSpeciation(tree2), estimateSpeciatio
 ```
 
     ##      Estimated True
-    ## [1,] 0.5585848  0.5
-    ## [2,] 0.9537667  1.0
-    ## [3,] 4.0484498  5.0
+    ## [1,] 0.5151949  0.5
+    ## [2,] 0.9099975  1.0
+    ## [3,] 5.6846979  5.0
 
 Now that you have a sense for the accuracy of our estimator, let’s
 simulate 100 phylogenetic trees with variable speciation rates and apply
@@ -356,7 +356,7 @@ fitCRBD(tree)
 ```
 
     ##        S        E 
-    ## 1.335349 1.337585
+    ## 1.634997 1.595090
 
 Let’s repeat our simulation procedure, but this time we will fit a model
 where extinction can be nonzero. We won’t display the plots here, but
