### 2026-01-24 f7fcb31
Update CITATION.cff
