+++ CITATION.cff
@@ -26,7 +26,7 @@ authors:
     orcid: https://orcid.org/0000-0002-7499-8251
     affiliation: Museum of Zoology and Department of Ecology and Evolutionary Biology, University of Michigan, Ann Arbor, MI, USA
 
-version: 1.0.1
+version: 1.0.23
 
 license: GPL-3.0
 
