+++ README.md
@@ -3,7 +3,7 @@ Instructions on EvolVEd, R, RStudio, RMarkdown
 Matheus Januario, Jenniefer Auler, Andressa Viol, and Daniel Rabosky
 Nov/2024
 
-[![DOI](https://zenodo.org/badge/530684084.svg)](https://doi.org/10.5281/zenodo.18203397)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.18203397.svg)](https://doi.org/10.5281/zenodo.18203397)
 
 - [EvolVEd](#evolved)
   - [Installing EvolVEd](#installing-evolved)
