+++ articles/birthdeath_phylogenies.md
@@ -230,9 +230,9 @@ matrix(c(estimateSpeciation(tree1), estimateSpeciation(tree2), estimateSpeciatio
 ```
 
     ##      Estimated True
-    ## [1,] 0.4769723  0.5
-    ## [2,] 1.0779045  1.0
-    ## [3,] 4.6338350  5.0
+    ## [1,] 0.5081707  0.5
+    ## [2,] 0.9795163  1.0
+    ## [3,] 4.7043952  5.0
 
 Now that you have a sense for the accuracy of our estimator, let’s
 simulate 100 phylogenetic trees with variable speciation rates and apply
@@ -356,7 +356,7 @@ fitCRBD(tree)
 ```
 
     ##        S        E 
-    ## 1.561232 1.622935
+    ## 1.190664 1.228156
 
 Let’s repeat our simulation procedure, but this time we will fit a model
 where extinction can be nonzero. We won’t display the plots here, but
