### 2026-01-24 #300 @mjanuario
Hi @ethanwhite , apologies for the delay. I’ve manually uploaded the final version of the package to Zenodo, and everything now matches (files, names, and version). DOI is https://doi.org/10.5281/zenodo.18203397

I tried using the GitHub–Zenodo integration first, but manual upload turned out to be simpler.
