### 2026-01-26 #300 @ethanwhite
Thanks @mjanuario! Everything looks good on my end. It's been a pleasure working with you on this!

I'll now recommend acceptance at which point this will go to the EIC for a final set of checks prior to publication.
