### 2026-01-26 #300 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
