### 2026-02-09 #300 @ethanwhite
No worries @mjanuario. The EIC will do that once they've completed their final checks and everything looks good to them.
