### 2026-02-11 #300 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
